% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics-area_gain_loss.R
\name{array2df}
\alias{array2df}
\title{Reshape array to data.frame}
\usage{
array2df(a, clim.scen, threshold, model, location)
}
\arguments{
\item{a}{array}

\item{clim.scen}{Climate scenario}

\item{threshold}{threshold}

\item{model}{model}

\item{location}{location}
}
\description{
Reshape array to data.frame
}
\keyword{internal}
