% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{bffr.b}
\alias{bffr.b}
\title{Create buffer based on species polygons}
\usage{
bffr.b(occ_polys, bffr.width = NULL, mult = 0.2, quadsegs = 100,
  o.path = "occ_poly", crs.set = NULL, plot = T)
}
\arguments{
\item{occ_polys}{list of SpatialPolygons, usually obj returned from poly.c.batch()}

\item{bffr.width}{Buffer width. See 'width' of ?rgeos::gBuffer}

\item{mult}{How much expand bffr.width}

\item{quadsegs}{see ?rgeos::gBuffer}

\item{o.path}{Output path}

\item{crs.set}{crs value used ???}

\item{plot}{Boolean, to draw plots or not}
}
\value{
A named list of SpatialPolygons
}
\description{
Create buffer based on species polygons
}
\examples{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"), "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ_polys <- poly.c.batch(spp.occ.list, o.path="occ_poly")
occ_b <- bffr.b(occ_polys, bffr.width=1.5)
}
