% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{bffr.batch}
\alias{bffr.batch}
\title{Create buffer based on species polygons}
\usage{
bffr.batch(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This funcion will create a buffer using species polygons. Buffer width may be manually specified or
calculated based on the extent of the SpatialPolygons object (i.e. mean of latitudinal and longitudinal extent).
If width is calculated based on the extent of the SpatialPolygons object, it can be adjusted (enlarged or reduced)
using 'mult' argument.
}
