% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{bind.shp}
\alias{bind.shp}
\title{bind list of SpatialPolygons into a single SpatialPolygon}
\usage{
bind.shp(files, sp.nm = "sp", o.path = "occ_poly", crs.set = NULL)
}
\arguments{
\item{files}{??}

\item{sp.nm}{name used to save file}

\item{o.path}{Output path}

\item{crs.set}{crs value used ???}
}
\value{
shapefile with binded polygons
}
\description{
bind list of SpatialPolygons into a single SpatialPolygon
}
