% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.f.data.prep.R
\name{bind_poly}
\alias{bind_poly}
\title{Bind list of SpatialPolygons into a single SpatialPolygons object}
\usage{
bind_poly(
  occ.polys,
  sp.nm = "species",
  save = TRUE,
  crs.set = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{occ.polys}{list of SpatialPolygons to bind}

\item{sp.nm}{Species name, used on saving shapefile}

\item{save}{Should save polygons on disk?}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\value{
shapefile with binded polygons
}
\description{
This function will bind a list of splitted polygons (from split_poly()) into a single SpatialPolygons object.
}
\seealso{
\code{\link{set_calibarea_b}}, \code{\link{set_calibarea}}, \code{\link{split_poly}}, \code{\link[NbClust]{NbClust}}
}
\keyword{internal}
