% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.f.data.prep.R
\name{buffer_b}
\alias{buffer_b}
\title{Create buffer based on species polygons}
\usage{
buffer_b(
  occ.polys,
  width = NULL,
  mult = 0.2,
  numCores = 1,
  crs.set = NULL,
  plot = F
)
}
\arguments{
\item{occ.polys}{list of SpatialPolygons, usually obj returned from set_calibarea_b()}

\item{width}{numeric > 0. Unit is meter if \code{x} has a longitude/latitude CRS, or mapunits in other cases}

\item{mult}{How much expand width}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}

\item{plot}{logical. Plot results or not?}
}
\value{
A named list of SpatialPolygons
}
\description{
This funcion will create a buffer using species polygons. Buffer width may be manually specified or
calculated based on the extent of the SpatialPolygons object (i.e. mean of latitudinal and longitudinal extent).
If width is calculated based on the extent of the SpatialPolygons object, it can be adjusted (enlarged or reduced)
using 'mult' argument.
}
\examples{
\dontrun{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- set_calibarea_b(spp.occ.list)
occ.b <- buffer_b(occ.polys, width=150000)
}
}
\seealso{
\code{\link[raster]{buffer}}, \code{\link{set_calibarea_b}}
}
