% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.2.f.metrics - OR for ensemble.R
\name{calc_or_ensemble}
\alias{calc_or_ensemble}
\title{Compute Omission Rate for a species' AICc Averaged Model}
\usage{
calc_or_ensemble(mcm = mxnt.mdls.preds.cf[[1]],
  a.calib = occ.b.env[[1]], ORt = seq(0, 0.2, 0.05), userArgs = NULL,
  categoricals)
}
\arguments{
\item{mcm}{Objects returned by "mxntCalib", containing calibrated models.}

\item{a.calib}{Predictors (cropped environmental variables) for model tuning. Used in model calibration. Argument 'x' of dismo::maxent. Raster* object or SpatialGridDataFrame, containing grids with
predictor variables. These will be used to extract values from for the point locations. Can
also be a data.frame, in which case each column should be a predictor variable and each row
a presence or background record.}

\item{ORt}{Threshold value to be used to compute OR}

\item{userArgs}{ character vector; use this to pass other arguments (e.g., prevalence) to the `maxent` call.  Note that not all options are functional or relevant.}

\item{categoricals}{ Vector indicating which (if any) of the input environmental layers are categorical. }
}
\value{
Data frame with average and variance of OR values across partition groups of data
}
\description{
This function will compute the omission rate (OR) for a species' AICc Averaged Model
from a 'mcmp' object, based on the selected threshold value.
}
\seealso{
\code{\link{f.thr.batch}}
}
