% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics.R
\name{cFPA}
\alias{cFPA}
\title{Compute "Fractional predicted area" ('n of occupied pixels'/n)}
\usage{
cFPA(mtp.l, digits = 3)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' FPAs computed for each climatic scenario, threshold and model criteria
}
\description{
Compute "Fractional predicted area" ('n of occupied pixels'/total n) or ('area of occupied pixels'/total area)
}
\examples{
cFPA(mtp.l=mods.thrshld.lst)
}
\seealso{
\code{\link{cSArea}}, \code{\link{cVarCI}}, \code{\link{cOR}}, \code{\link{cRasterOverlap}}
}
