% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.accessory.functions.R
\name{check_install_pkg}
\alias{check_install_pkg}
\title{Check if package is installed and install if needed}
\usage{
check_install_pkg(pkg)
}
\arguments{
\item{pkg}{packages to checked for installation}
}
\description{
This function will check if a package is installed and install it if needed
}
\keyword{internal}
