% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.1.community_metrics.R
\name{comm_gr_trait}
\alias{comm_gr_trait}
\title{Group taxa by trait}
\usage{
comm_gr_trait(mtp.l, trait.df, trait.cols, threshold = NULL, model = NULL)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{trait.df}{Data frame containing columns with taxon names (taxon),
and categorical traits}

\item{trait.cols}{Character. Name of columns containing traits}

\item{threshold}{Name of threshold to be found within mtp.l
object (returned by \code{\link{thrshld_b}})}

\item{model}{Name of model to be found within mtp.l
object (returned by \code{\link{thrshld_b}})}
}
\description{
This function will group taxa by trait (trait.df) and return a list of
presence/absence (binary distribution maps) stacks for each trait
}
