% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.1.community_metrics.R
\name{comm_spatial_metric}
\alias{comm_spatial_metric}
\title{Compute spatial metrics for each trait grouped taxa}
\usage{
comm_spatial_metric(
  trait.gr,
  fun = sum,
  user.fun = NULL,
  result.name = "richness",
  numCores = 1,
  format = "raster",
  ...
)
}
\arguments{
\item{trait.gr}{Object returned by \code{\link{comm_gr_trait}}}

\item{fun}{function to be passed to \code{\link[raster]{calc}}}

\item{user.fun}{User specified function. Can be used to compute focal/window metrics}

\item{result.name}{Name of directory where results will be saved}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{format}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{...}{Additional arguments: 
  		
\code{datatype}: Character. Output data type (e.g. 'INT2S' or 'FLT4S'). See \code{\link[raster]{dataType}}. If no datatype is specified, 'FLT4S' is used, unless this default value was changed with \code{\link[raster]{rasterOptions}}
	
\code{overwrite}: Logical. If TRUE, "filename" will be overwritten if it exists
	
\code{progress}: Character. Set a value to show a progress bar. Valid values are "text" and "window".

\code{NAflag}: Numeric. To overwrite the default value used to represent \code{NA} in a file 
	
\code{bandorder}: Character. 'BIL', 'BIP', or 'BSQ'. For 'native' file formats only.
For some other formats you can use the 'options' argument (see below)
	
\code{options}: Character. File format specific GDAL options. E.g., when
writing a geotiff file you can use: \code{options=c("COMPRESS=NONE", "TFW=YES")}

You can use options=c("PROFILE=BASELINE") to create a plain  tif with no GeoTIFF tags. 
This can be useful when writing files to be read by applications intolerant of unrecognised tags. 

NetCDF files have the following additional, optional, arguments: \code{varname}, \code{varunit}, \code{longname}, \code{xname}, \code{yname}, \code{zname}, \code{zunit}

\code{prj}: Logical. If \code{TRUE}, the crs is written to a .prj file. This can be useful 
when writing to an ascii file or another file type that does not store the crs

\code{setStatistics}: logical. If \code{TRUE} (the default) the min and max cell values are written to file (if the format permits it)
}
}
\description{
This function will compute user defined spatial metrics
for each trait grouped taxa
}
