% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.accessory.functions.R
\name{compass_rose}
\alias{compass_rose}
\title{Draw north arrow in plotted map}
\usage{
compass_rose(
  x,
  y,
  size,
  bearing = 0,
  d = 0.8,
  draw.circle = F,
  fill = c("white", "black"),
  cex = 1,
  lwd = 1,
  ...
)
}
\arguments{
\item{x}{Position at x axis}

\item{y}{Position at y axis}

\item{size}{Compass size}

\item{bearing}{bearing direction in degrees}

\item{d}{letter distance from circle}

\item{draw.circle}{Logical. Draw circle around compass?}

\item{fill}{The color for filling polygons of both sides of the north arrow or compass rose.
The default, c("white", border), fills one side with white and the other side
with the color of the border. Two colors should be provided.}

\item{cex}{A numerical value giving the amount by which plotting text
and symbols should be magnified relative to the default.
See \code{\link[graphics]{par}} for details}

\item{lwd}{A numerical value giving the amount by which plotting text
and symbols should be magnified relative to the default.
See \code{\link[graphics]{par}} for details}

\item{...}{graphical parameters such as \code{xpd}, \code{lend},
    \code{ljoin} and \code{lmitre} can be given as arguments.}
}
\description{
This function will draw a compass rose in the plotted map
}
\references{
Tanimura et al. 2007. Auxiliary Cartographic Functions in R: North
Arrow, Scale Bar, and Label with a Leader Arrow. Journal of Statistical
Software. Volume 19, Code Snippet 1. https://www.jstatsoft.org/article/view/v019c01/v19c01.pdf
}
