% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.f.consensus_projections.R
\name{consensus_gr}
\alias{consensus_gr}
\title{Group climate scenarios for consensual projections}
\usage{
consensus_gr(groups, clim.scn.nms)
}
\arguments{
\item{groups}{list containing vectors of names for grouping projections.}

\item{clim.scn.nms}{Vector with names of climate scenarios}
}
\value{
A data.frame with three columns: climate scenario names, group names,
and group numbers
}
\description{
This function groups climate scenarios by supplied groups (e.g. GCMs,
RCPs, SSPs) for consensual projections.
User need to supply a list containing vectors of names for grouping projections
and names of climate scenarios.
Projections will be grouped by matching the character vectors in the list
against projection names
}
\examples{
\dontrun{
# see projection names
names(mxnt.mdls.preds.cf[[1]]$mxnt.preds)
# vector with projection names
clim.scn.nms <- c("CCSM4.2050.RCP45",  "MIROC.ESM.2050.RCP45", "MPI.ESM.LR.2050.RCP45",
                  "CCSM4.2070.RCP45",  "MIROC.ESM.2070.RCP45", "MPI.ESM.LR.2070.RCP45",
                  "CCSM4.2050.RCP85",  "MIROC.ESM.2050.RCP85", "MPI.ESM.LR.2050.RCP85",
                  "CCSM4.2070.RCP85",  "MIROC.ESM.2070.RCP85", "MPI.ESM.LR.2070.RCP85")
# create two vectors containing grouping codes
yr <- c(2050, 2070)
rcp <- c("RCP45", "RCP85")
# run
consensus_gr(groups = list(yr, rcp), )
}
}
\seealso{
\code{\link{consensus_scn}}, \code{\link{consensus_scn_b}}
}
