% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.f.consensus_projections.R
\name{consensus_scn_b}
\alias{consensus_scn_b}
\title{Create consensual projections of climatic scenarios for multiple species}
\usage{
consensus_scn_b(mcmp.l, groups, ref = NULL, save = T)
}
\arguments{
\item{mcmp.l}{Object returned by \code{\link{proj_mdl_b}}, containing a list of calibrated models
and model projections for each species.}

\item{ref}{Object returned by \code{\link{proj_mdl_b}}, containing a list of calibrated models
and model projections for each species.}
}
\value{
A 'mcmp.l' object. An object returned from function \code{\link{proj_mdl_b}},
 containing the consensual projections for each element (species) of the list
}
\description{
This function creates consensual projections of climatic scenarios
(e.g. GCMs, RCPs, SSPs) for multiple species.
User need to supply a 'mcmp.l' object (returned by \code{\link{proj_mdl_b}}),
and a list containing names for grouping projections.
Projections will be grouped by matching names in the list of projections
}
\examples{
\dontrun{
# see projection names
names(mxnt.mdls.preds.cf[[1]]$mxnt.preds)
# create two vectors containing grouping codes
yr <- c(2050, 2070)
rcp <- c("RCP45", "RCP85")
# run
mxnt.mdls.preds.cf <- consensus_scn_b(mcmp.l=mxnt.mdls.preds.cf, groups = list(yr, rcp))
}
}
\seealso{
\code{\link{consensus_scn}}
}
