% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.f.metrics.R
\name{cOR}
\alias{cOR}
\title{Compute "Omission Rate"}
\usage{
cOR(mtp.l, occ.l, clim.scn.nm = NULL, digits = 3)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{occ.l}{list of species occurrence data}

\item{clim.scn.nm}{name to locate climatic scenario from which Omission Rate will
be extracted. Usually the scenario used to calibrate maxent models}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' ORs computed for the selected (current) climatic scenario and
each threshold and model criteria
}
\description{
Compute "Omission Rate" of species occurence points for a climatic scenario (usually "current")
}
\examples{
cOR(mtp.l=mods.thrshld.lst, occ.l=occ.locs, "current")
}
\seealso{
\code{\link{cSArea}}, \code{\link{cVarCI}}, \code{\link{cFPA}}, \code{\link{cRasterOverlap}}
}
