% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2b.f.sel.env.R
\name{correlated}
\alias{correlated}
\title{Find highly correlated variables}
\usage{
correlated(corr.mat, cutoff = 0.9, names = F)
}
\arguments{
\item{corr.mat}{A correlation matrix}

\item{cutoff}{A numeric value for the pair-wise absolute correlation cutoff.}

\item{names}{Logical. Should variable names or indexes be returned?}
}
\value{
A vector of variable names (when names = TRUE) or variable/column indexes
(when names = F). If there are no correlations above the cutoff, all variable
names (or indexes) is returned
}
\description{
This function creates a correlation matrix for the layers of a raster brick/stack
and returns a brick containing the least correlated variables below the cutoff value.
}
\details{
The function finds variables with correlation above the cutoff and
sucessively picks up the variable with the largest number of pair-wise
correlations above the cutoff.
At each step, the variable is assigned to a group containing all variables
with at least one correlated variable in common. The variable
with the largest within group mean correlation is discarded.
}
\seealso{
\code{\link{select_vars_b}}, \code{\link{select_vars}}, \code{\link[caret]{findCorrelation}}
}
