% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.f.niche.metrics.R
\name{cRasterOverlap}
\alias{cRasterOverlap}
\title{Raster overlap between models selected using different criteria}
\usage{
cRasterOverlap(mcmp.l, scn.nm = "current", model.compare = 1)
}
\arguments{
\item{mcmp.l}{Object returned by "mxntProjB", containing a list of calibrated models
and model projections for each species.}

\item{scn.nm}{name of climatic scenarios to compute overlap.}

\item{model.compare}{Reference model to be compared (AvgAIC, LowAIC, avg.test.orMTP,
avg.test.or10pct, avg.test.AUC.MTP, avg.test.AUC10pct)}
}
\value{
A list of matrices containing the three metrics (I, D, and Spearman rank correlation)
for each climatic scenario
}
\description{
Measures overlap between two ENMs. Used to compare differences among model selection criteria.
See ?ENMTools::raster.overlap for details.
}
\examples{
mxnt.mdls.ovlp <- cRasterOverlap(mxnt.mdls.preds.pf, scn.nm="current", 1)
}
