% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{cut.env}
\alias{cut.env}
\title{Crop environmental variables for each species}
\usage{
\method{cut}{env}(occ_b, env_uncut)
}
\arguments{
\item{occ_b}{polygon, usually a buffer}

\item{env_uncut}{raster brick or stack to be cropped}
}
\value{
list [for each species] of cropped environmental variables. Details in ?raster::crop
}
\description{
Crop environmental variables for each species
}
\examples{
env_uncut <- brick("path/to/env")
occ_b_env <- cut.env(occ_b, env_uncut)
for(i in 1:length(occ_b_env)){
   plot(occ_b_env[[i]][[1]])
   plot(occ_b[[i]], add=T)
}
}
