% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.f.data.prep.R
\name{cut_calibarea_b}
\alias{cut_calibarea_b}
\title{Cut calibration area based on a list of SpatialPolygons
Use a list of SpatialPolygons to crop environmental variables for each species.}
\usage{
cut_calibarea_b(poly.l, env.uncut, numCores = 1)
}
\arguments{
\item{poly.l}{list of SpatialPolygons, usually returned from "buffer_b" function}

\item{env.uncut}{raster brick or stack to be cropped}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
list [one element for each species] of cropped environmental variables. Details in ?raster::crop
}
\description{
Cut calibration area based on a list of SpatialPolygons
Use a list of SpatialPolygons to crop environmental variables for each species.
}
\examples{
\dontrun{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- set_calibarea_b(spp.occ.list)
occ.b <- buffer_b(occ.polys, width=1.5)
env.uncut <- brick("path/to/env")
occ.b.env <- cut_calibarea_b(occ.b, env.uncut)
for(i in 1:length(occ.b.env)){
   plot(occ.b.env[[i]][[1]])
   plot(occ.b[[i]], add=TRUE)
}
}
}
\seealso{
\code{\link{load_calib_data_b}}, \code{\link[raster]{crop}}, \code{\link{buffer_b}}
}
