% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{cut_projarea_b}
\alias{cut_projarea_b}
\title{Cut area for projection based on a list of SpatialPolygons}
\usage{
cut_projarea_b(pred.polys, env.uncut, prj.nm = "")
}
\arguments{
\item{pred.polys}{list of SpatialPolygons (usually of based on species occ points)

#' @seealso \code{\link{set_projarea}}, \code{\link{set_projarea_b}},
\code{\link{cut_projarea}}, \code{\link{cut_projarea_mscn_b}},
\code{\link{cut_projarea_rst}}, \code{\link{cut_projarea_rst_b}}, \code{\link{cut_projarea_rst_mscn_b}}}

\item{env.uncut}{raster/brick of environmental variables to be cut}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}
}
\value{
named list of cropped raster or brick
}
\description{
This function is a wrapper for "cut_projarea". See ?cut_projarea
It works with a named list of pred.poly to delimit the projection area for each of the species.
}
\examples{
area.projection <- cut_projarea_b(pred.polys, env.uncut)
plot(area.projection[[1]][[1]])
}
