% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{cut_projarea_mscn_b}
\alias{cut_projarea_mscn_b}
\title{Cut multiple projection areas (climatic scenarios) for multiple species (list of SpatialPolygons)}
\usage{
cut_projarea_mscn_b(pred.polys, env.uncut.l, numCores = 1)
}
\arguments{
\item{pred.polys}{list of SpatialPolygons (usually of based on species occ points)

#' @seealso \code{\link{set_projarea}}, \code{\link{set_projarea_b}},
\code{\link{cut_projarea}}, \code{\link{cut_projarea_mscn_b}},
\code{\link{cut_projarea_rst}}, \code{\link{cut_projarea_rst_b}}, \code{\link{cut_projarea_rst_mscn_b}}}

\item{env.uncut.l}{list of raster/brick of environmental variables to be cut}

\item{numCores}{specify number of cores if aim run in parallel}
}
\value{
named list of cropped list of raster/brick of environmental variables
}
\description{
This function is a wrapper for "cut_projarea". See ?cut_projarea. This function delimits the projection area for
each of the species contained in the pred.polys named list and crops
multiple rasters/bricks (i.e. representing distinct climatic scenaries) based on the same criteria for each species.
}
\seealso{
\code{\link{set_projarea}}, \code{\link{set_projarea_b}},
\code{\link{cut_projarea}}, \code{\link{cut_projarea_b}},
\code{\link{cut_projarea_rst}}, \code{\link{cut_projarea_rst_b}}, \code{\link{cut_projarea_rst_mscn_b}}
}
