% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{cut_projarea_rst}
\alias{cut_projarea_rst}
\title{Cut a projection area based on a SpatialPolygon (e.g. Ecoregion)}
\usage{
cut_projarea_rst(
  area.p,
  env.uncut,
  mask = FALSE,
  sp.mask = NULL,
  prj.nm = "",
  sp.nm = "species"
)
}
\arguments{
\item{area.p}{SpatialPolygon to be used as reference to crop/mask environmental variables of all species}

\item{env.uncut}{raster/brick of environmental variables to be cut}

\item{mask}{should use area.p to "mask" or "crop" env.uncut? See ?raster::mask and ?raster::crop for details}

\item{sp.mask}{an object of class: SpatialPolygons or SpatialPolygonsDataFrame, used as additional mask to be
applied on the predictor variables (env.uncut)}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{sp.nm}{name (of species) to give to saved object}
}
\value{
environmental layers (raster/brick) cutted
}
\description{
This function will use a single SpatialPolygon to crop/mask raster/brick objects to be used on model projections.
}
\seealso{
\code{\link{set_projarea}}, \code{\link{set_projarea_b}},
\code{\link{cut_projarea}}, \code{\link{cut_projarea_b}}, \code{\link{cut_projarea_mscn_b}},
\code{\link{cut_projarea_rst_b}}, \code{\link{cut_projarea_rst_mscn_b}}
}
