% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{cut_projarea_rst_b}
\alias{cut_projarea_rst_b}
\title{Cut projection areas of multiple species based on a single SpatialPolygon object (e.g. Ecoregion)}
\usage{
cut_projarea_rst_b(
  area.p,
  env.uncut,
  occ.polys,
  mask = FALSE,
  prj.nm = "",
  sp.nm = "mult.spp"
)
}
\arguments{
\item{area.p}{SpatialPolygon to be used as reference to crop/mask environmental variables of all species}

\item{env.uncut}{raster/brick of environmental variables to be cut}

\item{occ.polys}{list of occ.poly SpatialPolygons.  see ?set_projarea for details.}

\item{mask}{Should mask raster? (i.e. only use area inside polygon. See ?raster::mask for details) or
use all spatial extent of area.p

#' @seealso \code{\link{set_projarea}}, \code{\link{set_projarea_b}},
\code{\link{cut_projarea}}, \code{\link{cut_projarea_b}}, \code{\link{cut_projarea_mscn_b}},
\code{\link{cut_projarea_rst}}, \code{\link{cut_projarea_rst_mscn_b}}}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{sp.nm}{name (of species) to give to saved object}
}
\value{
list of environmental layers (raster/brick) cutted
}
\description{
This function will use a single SpatialPolygon to crop/mask raster/brick objects to be used on model projections.
}
