% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics.R
\name{cVarCI}
\alias{cVarCI}
\title{Compute variable contribution and permutation importance}
\usage{
cVarCI(mcmp.l)
}
\arguments{
\item{mcmp.l}{Object returned by "mxntProjB", containing a list of calibrated models
and model projections for each species.}
}
\value{
List of arrays containing variable contribution and importance for each species
}
\description{
Compute variable contribution and importance for each model
}
\examples{
cVarCI(mcmp.l = mxnt.mdls.preds.lst)
}
\seealso{
\code{\link[dismo]{maxent}}, \code{\link{cSArea}}, \code{\link{cOR}}, \code{\link{cFPA}}, \code{\link{cRasterOverlap}}
}
