% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1b.f.occ.thin.R
\name{e_thin_algorithm}
\alias{e_thin_algorithm}
\title{Thinning of occurrence records using environmental information}
\usage{
e_thin_algorithm(data, bins = 20)
}
\arguments{
\item{data}{A matrix or data.frame of environmental data at occurrence records.}

\item{bins}{Number of bins to divide the environmental range of
each variable.}
}
\value{
A vector containing the index (row numbers) of selected records.
}
\description{
The function thins spatial points to minimize biases in the environmental space.
It divides de range of variables into a desired number of @param bins and selects points
closer to the bin center.
}
