% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.accessory.functions.R
\name{empty2zero}
\alias{empty2zero}
\title{Check if element is empty}
\usage{
empty2zero(x)
}
\arguments{
\item{x}{value to be checked}
}
\description{
This function will check if any element is empty
}
\keyword{internal}
