% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{ENMevaluate.batch}
\alias{ENMevaluate.batch}
\title{Tuning and evaluation of ENMs with Maxent for several species using ENMeval}
\usage{
ENMevaluate.batch(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function is a wrapper for ENMeval::ENMevaluate. See ?ENMeval::ENMevaluate for details
It works with a named list of species occurrence data (occ.l) and a list of
cropped environmental variables (a.calib.l) for model tuning.
}
