% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{ENMwizard-deprecated-defunct}
\alias{ENMwizard-deprecated-defunct}
\alias{poly.c}
\alias{poly.c.batch}
\alias{bffr.batch}
\alias{env.cut}
\alias{thin.batch}
\alias{loadTocc}
\alias{sel_env_b}
\alias{sel_env}
\alias{pred.a.poly}
\alias{pred.a.poly.batch}
\alias{pred.a}
\alias{pred.a.batch}
\alias{pred.a.batch.mscn}
\alias{pred.a.rst}
\alias{pred.a.batch.rst}
\alias{pred.a.batch.rst.mscn}
\alias{ENMevaluate.batch}
\alias{ENMevaluate.l.opt}
\alias{f.args}
\alias{mxnt.c}
\alias{mxnt.c.batch}
\alias{mxnt.p}
\alias{mxnt.p.batch}
\alias{mxnt.p.batch.mscn}
\alias{f.thr}
\alias{f.thr.batch}
\alias{plotMdlDiff}
\alias{plotScnDiff}
\alias{f.area.occ.mscn}
\alias{f.var.ci}
\alias{f.FPA}
\alias{f.OR}
\alias{f.raster.overlap.mscn}
\title{Deprecated & Defunct Functions}
\usage{
poly.c(...)

poly.c.batch(...)

bffr.batch(...)

env.cut(...)

thin.batch(...)

loadTocc(...)

sel_env_b(...)

sel_env(...)

pred.a.poly(...)

pred.a.poly.batch(...)

pred.a(...)

pred.a.batch(...)

pred.a.batch.mscn(...)

pred.a.rst(...)

pred.a.batch.rst(...)

pred.a.batch.rst.mscn(...)

ENMevaluate.batch(...)

ENMevaluate.l.opt(...)

f.args(...)

mxnt.c(...)

mxnt.c.batch(...)

mxnt.p(...)

mxnt.p.batch(...)

mxnt.p.batch.mscn(...)

f.thr(...)

f.thr.batch(...)

plotMdlDiff(...)

plotScnDiff(...)

f.area.occ.mscn(...)

f.var.ci(...)

f.FPA(...)

f.OR(...)

f.raster.overlap.mscn(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
These functions are Deprecated or Defunct in this release of ENMwizard,
Deprecated functions will be marked as Defunct and removed in a future version.
Defunct functions were already removed from this version.

'poly.c' is defunct. Use 'set_calibarea' instead.

'poly.c.batch' is defunct. Use 'set_calibarea_b' instead.

'bffr.batch' is defunct. Use 'buffer_b' instead.

'env.cut' is defunct. Use 'cut_calibarea_b' instead.

'thin.batch' is defunct. Use 'thin_b' instead.

'loadTocc' is defunct. Use 'load_thin_occ' instead.

'sel_env_b' is defunct. Use 'select_vars_b' instead.

'sel_env' is defunct. Use 'select_vars' instead.

'pred.a.poly' is defunct. Use 'set_calibarea' instead.

'pred.a.poly.batch' is defunct. Use 'set_projarea_b' instead.

'pred.a' is defunct. Use 'cut_projarea_b' instead.

'pred.a.batch' is defunct. Use 'cut_projarea_b' instead.

'pred.a.batch.mscn' is defunct. Use 'cut_projarea_mscn_b' instead.

'pred.a.rst' is defunct. Use 'cut_projarea_rst' instead.

'pred.a.batch.rst' is defunct. Use 'cut_projarea_rst_b' instead.

'pred.a.batch.rst.mscn' is defunct. Use 'cut_projarea_rst_mscn_b' instead.

'ENMevaluate.batch' is defunct. Use 'ENMevaluate_b' instead.

'ENMevaluate.l.opt' is defunct. Use 'optENMevalObjL' instead.

'f.args' is defunct. Use 'mod_sel' instead.

'mxnt.c' is defunct. Use 'calib_mdl' instead.

'mxnt.c.batch' is defunct. Use 'calib_mdl_b' instead.

'mxnt.p' is defunct. Use 'proj_mdl' instead.

'mxnt.p.batch' is defunct. Use 'proj_mdl' instead.

'mxnt.p.batch.mscn' is defunct. Use 'proj_mdl_b' instead.

'f.thr' is defunct. Use 'thrshld' instead.

'f.thr.batch' is defunct. Use 'thrshld_b' instead.

'plotMdlDiff' is defunct. Use 'plot_mdl_diff_b' instead.

'plotScnDiff' is defunct. Use 'plot_scn_diff_b' instead.

'f.area.occ.mscn' is defunct. Use 'get_tsa_b' or 'get_tsa' instead.

'f.var.ci' is defunct. Use 'get_cont_permimport_b' or 'get_cont_permimport' instead.

'f.FPA' is defunct. Use 'get_fpa_b' or 'get_fpa' instead.

'f.OR' is defunct. Use 'get_OR' or 'get_or_ensemble' or 'get_or_ensemble_b' instead.

'f.raster.overlap.mscn' is defunct. Use 'raster_overlap_b' instead.
}
\keyword{internal}
