% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.f.mscn.projection.R
\name{ensemble_projs}
\alias{ensemble_projs}
\title{Ensemble models (AvgAIC, WAAUC, EBPM, ESORIC)}
\usage{
ensemble_projs(
  mcm,
  filename.or.auc.laic,
  a.proj,
  path.mdls,
  outpt,
  pred.nm,
  format
)
}
\arguments{
\item{mcm}{Objects returned by "calib_mdl", containing calibrated models.}

\item{a.proj}{A Raster* object or a data.frame where models will be projected. Argument 'x' of dismo::predict}

\item{pred.nm}{Character. Prefix to add to projection name (e.g. "current", "fut", or "past")}
}
\value{
A list containing all the items returned from function "calib_mdl", plus the projection specified in a.proj.
Each projection is a raster stack containing model projections ('mxnt.preds'), where each layer is a projection based on
a specific model selection criteria (i.e. AvgAIC, LowAIC, avg.test.orMTP, avg.test.or10pct, avg.test.AUC.MTP, avg.test.AUC10pct)
}
\description{
This function will read an object returned by "calib_mdl", read the calibrated models and project into
new areas/climatic scenarios. These new projections will be returned together with (appended to)
the original object.
}
\seealso{
\code{\link{mod_sel}}, \code{\link{calib_mdl}}, \code{\link{calib_mdl_b}}, \code{\link[dismo]{maxent}},
\code{\link[ENMeval]{ENMevaluate}}, \code{\link{proj_mdl_b}}
}
\keyword{internal}
