% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.f.mscn.projection.R
\name{EnsembleProjs}
\alias{EnsembleProjs}
\title{Ensemble models (AvgAIC, WAAUC, EBPM)}
\usage{
EnsembleProjs(mcm, filename.or.auc.laic, path.mdls, outpt, pred.nm, formt)
}
\arguments{
\item{mcm}{Objects returned by "mxntCalib", containing calibrated models.}

\item{pred.nm}{Character. Prefix to add to projection name (e.g. "current", "fut", or "past")}

\item{a.proj}{A Raster* object or a data.frame where models will be projected. Argument 'x' of dismo::predict}
}
\value{
A list containing all the items returned from function "mxntCalib", plus the projection specified in a.proj.
Each projection is a raster stack containing model projections ('mxnt.preds'), where each layer is a projection based on
a specific model selection criteria (i.e. AvgAIC, LowAIC, avg.test.orMTP, avg.test.or10pct, avg.test.AUC.MTP, avg.test.AUC10pct)
}
\description{
This function will read an object returned by "mxntCalib", read the calibrated models and project into
new areas/climatic scenarios. These new projections will be returned together with (appended to)
the original object.
}
\seealso{
\code{\link{modSel}}, \code{\link{mxntCalib}}, \code{\link{mxntCalibB}}, \code{\link[dismo]{maxent}},
\code{\link[ENMeval]{ENMevaluate}}, \code{\link{mxntProjB}}
}
\keyword{internal}
