% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{env.cut}
\alias{env.cut}
\title{Crop environmental variables for each species}
\usage{
env.cut(occ.b, env.uncut)
}
\arguments{
\item{occ.b}{polygon, usually a buffer}

\item{env.uncut}{raster brick or stack to be cropped}
}
\value{
list [for each species] of cropped environmental variables. Details in ?raster::crop
}
\description{
Crop environmental variables for each species
}
\examples{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- poly.c.batch(spp.occ.list)
occ.b <- bffr.batch(occ.polys, bffr.width=1.5)
env.uncut <- brick("path/to/env")
occ.b.env <- env.cut(occ.b, env.uncut)
for(i in 1:length(occ.b.env)){
   plot(occ.b.env[[i]][[1]])
   plot(occ.b[[i]], add=T)
}
}
