% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1b.f.occ.thin.R
\name{env_thin_b}
\alias{env_thin_b}
\title{Thinning of occurrence records using environmental information}
\usage{
env_thin_b(p.lst, predictors.lst, bins = 20, plot = F)
}
\arguments{
\item{p.lst}{List of two column matrix or data.frame with point coordinates
or \code{\link[sp]{SpatialPoints}} of occurrence records.}

\item{predictors.lst}{List of \code{\link[raster]{Raster-class}} object of environmental
predictor variables.}

\item{bins}{Number of bins to divide each environmental variable.}

\item{plot}{Logical. Should results be plotted?}
}
\value{
A data.frame, matrix, sp or sf object containing the selected records.
}
\description{
The function thins spatial points to minimize biases in the environmental space.
It divides de range of variables into a desired number of @param bins and selects points
closer to the bin center.
}
\examples{
\dontrun{
file <- paste(system.file(package="dismo"), "/ex/bradypus.csv", sep="")
bradypus <- read.table(file, header=TRUE, sep=',')
coord <- bradypus [,2:3]

# Download data for present
dir.create("rasters")
predictors <- getData('worldclim', var='bio', res=10, path="rasters")

#cor(data)
#vars <- ENMwizard::select_vars(cutoff = 0.9, corr.mat = cor(data), names.only = T)[[1]]
#length(vars)
env_thin_b(list(coord), list(predictors))
}

}
