% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.f.data.prep.R
\name{envCut}
\alias{envCut}
\title{Cut calibration area based on a list of SpatialPolygons
Use a list of SpatialPolygons to crop environmental variables for each species.}
\usage{
envCut(occ.b, env.uncut, numCores = 1)
}
\arguments{
\item{occ.b}{list of SpatialPolygons, usually returned from "bffrB" function}

\item{env.uncut}{raster brick or stack to be cropped}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
list [one element for each species] of cropped environmental variables. Details in ?raster::crop
}
\description{
Cut calibration area based on a list of SpatialPolygons
Use a list of SpatialPolygons to crop environmental variables for each species.
}
\examples{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- polyCB(spp.occ.list)
occ.b <- bffrB(occ.polys, bffr.width=1.5)
env.uncut <- brick("path/to/env")
occ.b.env <- envCut(occ.b, env.uncut)
for(i in 1:length(occ.b.env)){
   plot(occ.b.env[[i]][[1]])
   plot(occ.b[[i]], add=TRUE)
}
}
\seealso{
\code{\link[raster]{crop}}, \code{\link{bffrB}}
}
