% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{f.area.occ.mscn}
\alias{f.area.occ.mscn}
\title{Plot differences in suitable areas between climatic scenarios}
\usage{
f.area.occ.mscn(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Plot differences between a selected climatic scenario and all other climatic scenarios for each species.
This function will plota and (optionally) save the figures on pdf files in the folder "Mdls.thrshld/figs".
}
\details{
Compute species' total suitable area

Compute total suitable area at multiple climatic scenario, threshold and model criteria.
}
