% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.f.tuning.final.R
\name{f.args}
\alias{f.args}
\title{Creating arguments for selected models}
\usage{
f.args(x, wAICsum = 0.99, save = "B", randomseed = F,
  responsecurves = T, arg1 = "noaddsamplestobackground",
  arg2 = "noautofeature")
}
\arguments{
\item{x}{object of class ENMevaluation}

\item{wAICsum}{cumulative sum of top ranked models for which arguments will be created}

\item{save}{should save args only ("A"), selected models only ("M") or both ("B")?}

\item{randomseed}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{responsecurves}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg1}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg2}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}
}
\value{
A vector of args (if save="A"), data.frame of selected models (if save="M") or
a list with both, args and selected models, (if save="B")
}
\description{
This function will read an object of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return the necessary arguments for final model calibration and predictions.
}
\examples{
ENMeval_res.lst <- ENMevaluate.batch(occ_locs, occ_b_env, parallel = T , numCores = 7)
f.args(ENMeval_res.lst[[1]]@results)
}
