% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{f.bffr}
\alias{f.bffr}
\title{Create buffer based on polygon}
\usage{
f.bffr(occ_polys, bffr.width = NULL, mult = 0.2, quadsegs = 100,
  o.path = "occ_poly", crs.set = NULL, plot = T)
}
\arguments{
\item{occ_polys}{Spatial polygon}

\item{bffr.width}{Buffer width. See 'width' of ?rgeos::gBuffer}

\item{mult}{How much expand bffr.width}

\item{quadsegs}{see ?rgeos::gBuffer}

\item{o.path}{Output path}

\item{crs.set}{crs value used ???}

\item{plot}{Boolean, to draw plots or not}
}
\value{
A named list of SpatialPolygons
}
\description{
Create buffer based on polygon
}
\examples{
occ_b <- f.bffr(occ_polys, bffr.width=1.5, crs.set=crs.set) #
}
