% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{f.bind.shp}
\alias{f.bind.shp}
\title{merge several shapefiles into a single one}
\usage{
f.bind.shp(files, sp.nm = "sp", o.path = "occ_poly", crs.set = NA)
}
\arguments{
\item{files}{??}

\item{sp.nm}{name used to save file}

\item{o.path}{Output path}

\item{crs.set}{crs value used ???}
}
\value{
shapefile with binded polygons
}
\description{
merge several shapefiles into a single one
}
