% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{f.FPA}
\alias{f.FPA}
\title{Compute variable contribution and permutation importance}
\usage{
f.FPA(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Compute variable contribution and importance for each model
}
\details{
Compute "Fractional predicted area" ('n of occupied pixels'/n)

Compute "Fractional predicted area" ('n of occupied pixels'/total n) or ('area of occupied pixels'/total area)
}
