% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{f.load_occ_T}
\alias{f.load_occ_T}
\title{Load occurrence data filtered using "thin.batch"}
\usage{
f.load_occ_T(occ.list.thin, from.disk = F, wtd = 1)
}
\arguments{
\item{occ.list.thin}{named list returned from "thin.batch"}

\item{from.disk}{boolean. Read from disk or from one of thinned datasets stored in 'occ.list.thin' obj}

\item{wtd}{: which thinned dataset?}
}
\value{
named list of thinned occurence data for each species in the list
}
\description{
Load occurrence data filtered using "thin.batch"
}
\examples{
occ_locs <- f.load_occ_T(thinned_dataset_batch)
}
