% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.f.tuning.final.R
\name{f.mxnt.mdl.pred}
\alias{f.mxnt.mdl.pred}
\title{Calibrating and predicting selected models}
\usage{
f.mxnt.mdl.pred(x, sp.nm, area_projection, occ_b_env, occ_locs,
  formt = "raster", pred.args = c("outputformat=cloglog", "doclamp=true",
  "pictures=true"), wAICsum = 0.99, randomseed = F, responsecurves = T,
  arg1 = "noaddsamplestobackground", arg2 = "noautofeature")
}
\arguments{
\item{x}{Slot "results" of object of class ENMevaluation}

\item{sp.nm}{Species name. Used to name the output folder}

\item{area_projection}{A Raster* object or a data.frame where models will be projected. Argument 'x' of dismo::predict}

\item{occ_b_env}{Predictors. Used in model calibration. Argument 'x' of dismo::maxent. Raster* object or SpatialGridDataFrame, containing grids with
predictor variables. These will be used to extract values from for the point locations. Can
also be a data.frame, in which case each column should be a predictor variable and each row
a presence or background record.}

\item{occ_locs}{Occurrence data. Argument 'p' of dismo::maxent. This can be a data.frame, matrix,
SpatialPoints* object, or a vector. If p is a data.frame or matrix it represents a set of point
locations; and it must have two columns with the first being the x-coordinate (longitude) and
the second the y-coordinate (latitude). Coordinates can also be specified with a SpatialPoints*
object
If x is a data.frame, p should be a vector with a length equal to nrow(x) and contain 0
(background) and 1 (presence) values, to indicate which records (rows) in data.frame x are
presence records, and which are background records}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{pred.args}{Charater. Argument 'args' of dismo::maxent. Additional argument that can be
passed to MaxEnt. See the MaxEnt help for more information. The R maxent function only uses the
arguments relevant to model fitting. There is no point in using args='outputformat=raw' when
*fitting* the model; but you can use arguments relevant for *prediction* when using the predict
function. Some other arguments do not apply at all to the R implementation. An example is
'outputfiletype', because the 'predict' function has its own 'filename' argument for that.}

\item{wAICsum}{cumulative sum of top ranked models for which arguments will be created}

\item{randomseed}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{responsecurves}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg1}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg2}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}
}
\value{
A list containing the models ('selected.mdls') used for model calibration and prediction,
calibrated maxent models ('mxnt.mdls'), arguments used for prediction/calibration ('pred.args'), and
a raster stack containing model projections ('mxnt.preds'), where each layer is a projection based on
a specific model selection criteria (i.e. AvgAICc, LowAICc, Mean.ORmin, Mean.OR10, Mean.AUC_min, Mean.AUC_10)
}
\description{
This function will read an object of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return selected maxent model calibrations and predictions.
}
