% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{f.OR}
\alias{f.OR}
\title{Compute "Fractional predicted area" ('n of occupied pixels'/n)}
\usage{
f.OR(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Compute "Fractional predicted area" ('n of occupied pixels'/total n) or ('area of occupied pixels'/total area)
}
\details{
Compute "Omission Rate"

Compute "Omission Rate" of species occurence points for a climatic scenario (usually "current")
}
