% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{f.plot.mxnt.preds.mscn}
\alias{f.plot.mxnt.preds.mscn}
\title{Plot differences (for multiple climatic scenarios) among model predictions selected from several criteria}
\usage{
f.plot.mxnt.preds.mscn(mmp.spl, mtp.spl, basemap = NULL, numCores = 1)
}
\arguments{
\item{mmp.spl}{List of stack or brick of predictions to apply the threshold}

\item{mtp.spl}{List of stack or brick of thresholded predictions}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
won't return any object. Will save pdf's with differences among model predictions (for multiple climatic scenarios)
}
\description{
Plot differences (for multiple climatic scenarios) among model predictions selected from several criteria
}
\examples{
f.plot.mxnt.preds.mscn(mxnt.mdls.preds.lst, mods.thrshld.lst, basemap=NewWorld)
f.plot.mxnt.preds.mscn(mxnt.mdls.preds.pf[1], mods.thrshld.lst[1], basemap=NewWorld)
}
