% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{f.plot.mxnt.preds}
\alias{f.plot.mxnt.preds}
\title{Plot differences among predictions of models selected from several criteria}
\usage{
f.plot.mxnt.preds(mmp.spl, mtp.spl, basemap = NULL)
}
\arguments{
\item{mmp.spl}{Object returned by "mxnt.p.batch.mscn", containing a list of calibrated models
and model projections for each species.}

\item{mtp.spl}{List of stack or brick of thresholded predictions}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}
}
\value{
won't return any object. Will save pdf's with differences among model predictions
}
\description{
Plot differences among predictions of models selected from several criteria.
This function will save the figures on pdf files in the folder "Mdls.thrshld/figs"
}
\examples{
f.plot.mxnt.preds(mxnt.mdls.preds.lst, mods.thrshld.lst, basemap=NewWorld)
}
