% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{f.plot.scn.diff}
\alias{f.plot.scn.diff}
\title{Plot differences between climatic scenarios}
\usage{
f.plot.scn.diff(mmp.spl, mtp.spl, sel.clim.scn = "current", basemap = NULL,
  save = F, numCores = 1)
}
\arguments{
\item{mmp.spl}{Object returned by "mxnt.p.batch.mscn", containing a list of calibrated models
and model projections for each species.}

\item{mtp.spl}{List of stack or brick of thresholded predictions}

\item{sel.clim.scn}{Selected climatic scenario to compare with all others. Usually "current" one.}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
won't return any object. Will save pdf's with differences among model predictions (for multiple climatic scenarios)
}
\description{
Plot differences between a selected climatic scenario and all other climatic scenarios.
This function will plota and (optionally) save the figures on pdf files in the folder "Mdls.thrshld/figs"
}
\examples{
f.plot.scn.diff(mxnt.mdls.preds.cf, mods.thrshld.lst)
}
