% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{f.plotScn}
\alias{f.plotScn}
\title{internal function for \code{\link{plotMdlDiff}}, \code{\link{plotScnDiffB}}}
\usage{
f.plotScn(m, mtp, ref.scn = "current", basemap, thrshld.path,
  thrshld.nms.mod, save, msnm, msr, scnm, scr)
}
\arguments{
\item{m}{Index of model selection criteria to be plotted}

\item{mtp}{Stack or brick of thresholded predictions}

\item{ref.scn}{Selected climatic scenario to compare with all others. Usually "current" one.}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}

\item{thrshld.path}{path to threshold projections}

\item{thrshld.nms.mod}{names of threshold models}

\item{save}{Logical. If TRUE will save plots in pdf.}

\item{msnm}{Character vector. Short names of model selection criteria to be replaced. Same as given in
Model names (ex. AUC10, AUCmtp, OR10, ORmtp)}

\item{msr}{Character vector. Long names of model selection criteria to replace the short names. Must
be in same order of argument 'msnm'}

\item{scnm}{Character vector. Short names of climatic scenarios to be replaced. Ex. "cc26bi70", "cc45bi70"}

\item{scr}{Character vector. Long names of climatic scenarios to replace the short names. Must
be in same order of argument 'scnm'. Ex. "2070-CCSM4-rcp2.6", "2070-CCSM4-rcp4.5"}
}
\description{
internal function for \code{\link{plotMdlDiff}}, \code{\link{plotScnDiffB}}
}
\keyword{internal}
