% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{f.poly.batch}
\alias{f.poly.batch}
\title{Create occ polygon for several species}
\usage{
f.poly.batch(spp.occ.list, o.path = NULL, crs.set = NA, convex = T,
  alpha = 10)
}
\arguments{
\item{spp.occ.list}{A named list of spatial data.frame of species occurence points}

\item{o.path}{Output path}

\item{crs.set}{crs value used ???}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{see ?alphahull::ashape}
}
\value{
A named list of spatial polygons built using coordinates
}
\description{
Create occ polygon for several species
}
\examples{
occ_polys <- f.poly.batch(spp.occ.list, o.path="occ_poly", convex=T, alpha=10)
}
