% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{f.poly}
\alias{f.poly}
\title{Create occ polygon to crop rasters prior to modelling}
\usage{
f.poly(occ.spdf, o.path = NULL, lr.nm = "sp_nm", convex = T, alpha = 10,
  crs.set = NA)
}
\arguments{
\item{occ.spdf}{A spatial data.frame of coordinates, usually species occurence coordinates}

\item{o.path}{Output path}

\item{lr.nm}{Polygon output name}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{see ?alphahull::ashape}

\item{crs.set}{crs value used ???}
}
\value{
spatial polygon built using coordinates
}
\description{
Create occ polygon to crop rasters prior to modelling
}
\examples{
occ_poly <- f.poly(occ.spdf, o.path="occ_poly", lr.nm="occ_poly")
plot(occ_poly)
}
