% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{f.poly.splt}
\alias{f.poly.splt}
\title{create N polygons for a species (whenever distribution seems disjoint) and save in a single .shp}
\usage{
f.poly.splt(spp.occ, k = 2, convex = T, alpha = 10, sp.nm = "sp1",
  o.path = "occ_poly", crs.set = NA)
}
\arguments{
\item{spp.occ}{species occurence coordinates}

\item{k}{number of polygons to create based on coordinates}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{see ?alphahull::ashape}

\item{sp.nm}{name used to save file}

\item{o.path}{Output path}

\item{crs.set}{crs value used ???}
}
\value{
spatial polygons built using coordinates
}
\description{
create N polygons for a species (whenever distribution seems disjoint) and save in a single .shp
}
\examples{
occ_polys$Bvarieg <- f.poly.splt(spp.occ = Bvarieg.occ, k=5, convex=T, alpha=10, sp.nm = "Bvarieg", o.path = "occ_poly", crs.set = crs.set)
}
