% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.f.niche.metrics.R
\name{f.raster.overlap.mscn}
\alias{f.raster.overlap.mscn}
\title{Raster overlap between models selected using different criteria}
\usage{
f.raster.overlap.mscn(mmp.spl, scn.nm = "current", model.compare = 1)
}
\arguments{
\item{mmp.spl}{Object returned by "mxnt.p.batch.mscn", containing a list of calibrated models
and model projections for each species.}

\item{scn.nm}{name of climatic scenarios to compute overlap.}

\item{model.compare}{Reference model to be compared (AvgAICc, LowAICc, Mean.ORmin,
Mean.OR10, Mean.AUCmin, Mean.AUC10)}
}
\value{
A list of matrices containing the three metrics (I, D, and Spearman rank correlation)
for each climatic scenario
}
\description{
Measures overlap between two ENMs. Used to compare differences among model selection criteria.
See ?ENMTools::raster.overlap for details.
}
\examples{
mxnt.mdls.ovlp <- f.raster.overlap.mscn(mxnt.mdls.preds.pf, scn.nm="current", 1)
#' mxnt.mdls.ovlp <- f.raster.overlap.mscn(mxnt.mdls.preds.pf, scn.nm=c("futAC5085", "futAC7085"), 3)
}
