% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.f.threshold.R
\name{f.thr.batch}
\alias{f.thr.batch}
\title{Apply threshold for all predictions}
\usage{
f.thr.batch(mmp.spl, thrshld.i = 4:6, numCores = 1)
}
\arguments{
\item{mmp.spl}{List of stack or brick of predictions to apply the threshold}

\item{thrshld.i}{List of threshold criteria to be applied}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
List of stack or brick of thresholded predictions
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
mods.thrshld.lst <- f.thr.batch(mxnt.mdls.preds.pf)
}
