% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.f.threshold.R
\name{f.thr}
\alias{f.thr}
\title{Apply threshold for a prediction}
\usage{
f.thr(mmp.spi, scn.nm = "", thrshld.i = 4:6, path.mdls = NULL)
}
\arguments{
\item{mmp.spi}{Species "i" of a object returned by "mxnt.p.batch.mscn", containing a list of
calibrated models and model projections for each species}

\item{scn.nm}{Name of climatic scenario to be looked for}

\item{thrshld.i}{List of threshold criteria to be applied}

\item{path.mdls}{Path where thresholded rasters will be saved}
}
\value{
Stack or brick of thresholded predictions
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
mods.thrshld <- f.thr(mxnt.mdls.preds, thrshld.i = 4:6, pred.args, path.mdls)
plot(mods.thrshld[[1]][[2]]) # continuous
plot(mods.thrshld[[2]][[2]]) # binary
}
