% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{f.var.ci}
\alias{f.var.ci}
\title{Compute species' total suitable area}
\usage{
f.var.ci(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Compute total suitable area at multiple climatic scenario, threshold and model criteria.
}
\details{
Compute variable contribution and permutation importance

Compute variable contribution and importance for each model
}
