% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.accessory.functions.R
\name{find_repeated_characters}
\alias{find_repeated_characters}
\title{Find repeated characters in  a character vector}
\usage{
find_repeated_characters(x)
}
\arguments{
\item{x}{Character vector}
}
\description{
This function will take the shorter element and the characters repeated
in the other elements of the character vector
}
\author{
Neander M. Heming
}
