% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics.R
\name{get_fpa}
\alias{get_fpa}
\title{Compute "Fractional predicted area" ('n of occupied pixels'/n)}
\usage{
get_fpa(mtp, digits)
}
\arguments{
\item{mtp}{Stack or brick of thresholded predictions}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' FPAs computed for each climatic scenario, threshold and model criteria
}
\description{
Compute "Fractional predicted area" ('n of occupied pixels'/total n) or ('area of occupied pixels'/total area)
}
\examples{
\dontrun{
get_fpa(mtp.l=mods.thrshld.lst)
}
}
\seealso{
\code{\link{get_fpa_b}}, \code{\link{get_tsa}}, \code{\link{get_cont_permimport}}
}
