% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics.R
\name{get_fpa_b}
\alias{get_fpa_b}
\title{Compute "Fractional predicted area" ('n of occupied pixels'/n)}
\usage{
get_fpa_b(mtp.l, digits = 3)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' FPAs computed for each climatic scenario, threshold and model criteria
}
\description{
Compute "Fractional predicted area" ('n of occupied pixels'/total n) or ('area of occupied pixels'/total area)
}
\examples{
\dontrun{
get_fpa_b(mtp.l=mods.thrshld.lst)
}
}
\seealso{
\code{\link{get_fpa}}, \code{\link{get_tsa}}, \code{\link{get_cont_permimport}}

\code{\link{get_tsa_b}}, \code{\link{get_cont_permimport_b}}
}
