% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.4.f.metrics.occ.suit.R
\name{get_occ_suit}
\alias{get_occ_suit}
\title{Extract suitability values from all projections for known species occurrences
of a 'mcmp' object}
\usage{
get_occ_suit(mcmp, occ = NULL, ref = "current", t.all = F)
}
\arguments{
\item{mcmp}{An object returned by \code{\link{proj_mdl}}, containing calibrated models and model
projections for each species; or species "i" of a object returned by "proj_mdl_b", containing a list of
calibrated models and model projections for each species}

\item{occ}{Occurrence data. Argument 'p' of dismo::maxent. This can be a data.frame, matrix,
SpatialPoints* object, or a vector. If p is a data.frame or matrix it represents a set of point
locations; and it must have two columns with the first being the x-coordinate (longitude) and
the second the y-coordinate (latitude). Coordinates can also be specified with a SpatialPoints*
object
If x is a data.frame, p should be a vector with a length equal to nrow(x) and contain 0
(background) and 1 (presence) values, to indicate which records (rows) in data.frame x are
presence records, and which are background records}

\item{ref}{Character. Name of reference projection (i.e. the one used for
calibration and that will not be averaged with any other)}

\item{t.all}{logical. Should threshold be applied on individual and consensus projections?
Default is FALSE. Ignored if consensus projections are not found.}
}
\description{
This function will extract the suitability values from all projections of a 'mcmp'
object using either the species occurrences used for niche modeling or an user
supplied object with species occurrences.
}
\examples{
\dontrun{
get_occ_suit(mcmp=mxnt.mdls.preds)
}
}
