% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.4.f.metrics.occ.suit.R
\name{get_occ_suit_b}
\alias{get_occ_suit_b}
\title{Extract suitability values from all projections for known species occurrences from
a 'mcmp.l' object}
\usage{
get_occ_suit_b(mcmp.l, occ.l = NULL, ref = "current", t.all = F)
}
\arguments{
\item{mcmp.l}{Object returned by \code{\link{proj_mdl_b}}, containing a list of calibrated models
and model projections for each species.}

\item{occ.l}{List of occurence data. See argument "occ" in calib_mdl.}

\item{ref}{Character. Name of reference projection (i.e. the one used for
calibration and that will not be averaged with any other)}

\item{t.all}{logical. Should threshold be applied on individual and consensus projections?
Default is FALSE. Ignored if consensus projections are not found.}
}
\description{
This function will extract the suitability values from all projections of each species
in a 'mcmp.l' object using either the species occurrences used for niche modeling or
a user supplied list of objects containing species occurrences.
}
\examples{
\dontrun{
get_occ_suit_b(mcmp.l=mxnt.mdls.preds.cf)
}
}
