% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics.R
\name{get_OR}
\alias{get_OR}
\title{Compute "Omission Rate"}
\usage{
get_OR(mtp.l, occ.l, clim.scn.nm = "current", digits = 3)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{occ.l}{list of species occurrence data}

\item{clim.scn.nm}{name to locate climatic scenario from which Omission Rate will
be extracted. Usually the scenario used to calibrate maxent models}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' ORs computed for the selected (current) climatic scenario and
each threshold and model criteria
}
\description{
Compute "Omission Rate" of species occurence points for a climatic scenario (usually "current")
}
\examples{
\dontrun{
get_OR(mtp.l=mods.thrshld.lst, occ.l=occ.locs)
}
}
\seealso{
\code{\link{get_tsa}}, \code{\link{get_cont_permimport}}, \code{\link{get_fpa}}
}
