% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.2.f.metrics.OR.ensemble.R
\name{get_or_ensemble}
\alias{get_or_ensemble}
\title{Compute Omission Rate for a species' ensembled model}
\usage{
get_or_ensemble(
  mcm,
  a.calib,
  ORt = seq(0, 0.2, 0.05),
  userArgs = NULL,
  categoricals,
  sp.nm = "species"
)
}
\arguments{
\item{mcm}{Objects returned by "calib_mdl", containing calibrated models.}

\item{a.calib}{Predictors (cropped environmental variables) for model tuning. Used in model calibration. Argument 'x' of dismo::maxent. Raster* object or SpatialGridDataFrame, containing grids with
predictor variables. These will be used to extract values from for the point locations. Can
also be a data.frame, in which case each column should be a predictor variable and each row
a presence or background record.}

\item{ORt}{Threshold value to be used to compute OR}

\item{userArgs}{ character vector; use this to pass other arguments (e.g., prevalence) to the `maxent` call.  Note that not all options are functional or relevant.}

\item{categoricals}{ Vector indicating which (if any) of the input environmental layers are categorical. }

\item{sp.nm}{Species name. Used to name the output folder}
}
\value{
Data frame with average and variance of OR values across partition groups of data
}
\description{
This function will compute the omission rate (OR) for a species' ensembled model
from a 'mcmp' object, based on the selected threshold value.
}
\seealso{
\code{\link{get_or_ensemble_b}}, \code{\link{get_tsa}}, \code{\link{get_cont_permimport}},
\code{\link{get_fpa}}, \code{\link{get_cont_permimport_b}}, \code{\link{get_fpa_b}}, \code{\link{get_tsa_b}}
}
