% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.2.f.metrics.OR.ensemble.R
\name{get_or_ensemble_b}
\alias{get_or_ensemble_b}
\title{Compute Omission Rate for a list of species' ensembled models}
\usage{
get_or_ensemble_b(mcm.l, a.calib.l, ORt = seq(0, 0.2, 0.05),
  userArgs = NULL, categoricals = NULL, numCores = 1)
}
\arguments{
\item{mcm.l}{A list of objects returned by "calib_mdl", containing calibrated models.}

\item{a.calib.l}{List of predictors (cropped environmental variables) for model tuning. Used in model calibration. Argument 'x' of dismo::maxent. Raster* object or SpatialGridDataFrame, containing grids with
predictor variables. These will be used to extract values from for the point locations. Can
also be a data.frame, in which case each column should be a predictor variable and each row
a presence or background record..}

\item{ORt}{Threshold value to be used to compute OR}

\item{userArgs}{ character vector; use this to pass other arguments (e.g., prevalence) to the `maxent` call.  Note that not all options are functional or relevant.}

\item{categoricals}{ Vector indicating which (if any) of the input environmental layers are categorical. }

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
Data frame with average and variance of OR values across partition groups of data
}
\description{
This function will compute the omission rate (OR) for each species' ensembled model
from a 'mcmp.l' object, based on the selected threshold value.
}
\seealso{
\code{\link{get_or_ensemble}}
}
