% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics-area_gain_loss.R
\name{get_rsa}
\alias{get_rsa}
\title{Compute changes in total suitable area}
\usage{
get_rsa(mrs, area.raster = NULL, digits = 2)
}
\arguments{
\item{mrs}{List containing stack or brick of range shifts. See \code{\link{range_shift}}}

\item{area.raster}{A raster containing the cell areas to be summed across
the suitable pixels. This allows summing areas of habitat when the pixel is
partially occupied with the habitat of interest.}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
An object of class "data.table" "data.frame" containing calculated areas of
 range shift between a selected climatic scenario and all other climatic scenarios
 for threshold, model and masked locations (if projections were masked, see \code{\link{mask_thr_projs_mscn_b}}).
}
\description{
Compute changes in total suitable area at multiple climatic scenarios, threshold and model criteria.
}
\examples{
\dontrun{
range_diff <- range_shift(mods.thrshld.lst$haddadus_binotatus)
spp_diff_a <- get_rsa(mrs=range_diff)
}
}
\seealso{
\code{\link[raster]{area}}, \code{\link{get_tsa_b}}, \code{\link{get_cont_permimport}}, \code{\link{get_fpa}},
\code{\link{get_cont_permimport_b}}, \code{\link{get_fpa_b}}
}
