% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics.R
\name{get_tsa_b}
\alias{get_tsa_b}
\title{Compute species' total suitable area}
\usage{
get_tsa_b(mtp.l, restrict = NULL, digits = 0)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{restrict}{a raster to select a region to compute area.}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
List of arrays containing species' total suitable areas for each climatic scenario, threshold and model criteria
}
\description{
Compute total suitable area at multiple climatic scenario, threshold and model criteria.
}
\examples{
\dontrun{
areas.occ.lst <- get_tsa_b(mtp.l=mods.thrshld.lst)
}
}
\seealso{
\code{\link[raster]{area}}, \code{\link{get_tsa}}, \code{\link{get_cont_permimport}},
\code{\link{get_fpa}}, \code{\link{get_cont_permimport_b}}, \code{\link{get_fpa_b}}
}
