% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.accessory.functions.R
\name{is_inst}
\alias{is_inst}
\title{Check if package is installed}
\usage{
is_inst(pkg)
}
\arguments{
\item{pkg}{packages to checked for installation}
}
\description{
This function will check if a package is installed
}
\keyword{internal}
