% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.f.data.prep.R
\name{load_calib_data_b}
\alias{load_calib_data_b}
\title{Load calibration areas already created}
\usage{
load_calib_data_b(poly.l, sel = F)
}
\arguments{
\item{poly.l}{list of SpatialPolygons, usually returned from "buffer_b" function}

\item{sel}{Logical. If TRUE, load selected (uncorrelated) variables,
if FALSE (default), load all variables. See \code{\link{select_vars_b}}
for details about selection of uncorrelated variables.}
}
\value{
list [one element for each species] of cropped environmental variables. Details in ?raster::crop
}
\description{
Use a named list to load cropped environmental variables
for each species.
}
\seealso{
\code{\link{cut_calibarea_b}}, \code{\link{select_vars_b}}
}
