% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1b.f.occ.thin.R
\name{load_env_thin_occ}
\alias{load_env_thin_occ}
\title{Load filtered occurrence data}
\usage{
load_env_thin_occ(p.lst, long.col = NULL, lat.col = NULL)
}
\arguments{
\item{p.lst}{named list returned from \code{\link{env_thin_b}}}

\item{long.col}{Name of columns that contain coordinates (latitude and longitude)}

\item{lat.col}{Name of columns that contain coordinates (latitude and longitude)}
}
\value{
named list of thinned occurence data for each species in the list
}
\description{
Load filtered occurrence data from object returned by \code{\link{env_thin_b}}
}
\examples{
\dontrun{
occ.locs <- load_env_thin_occ(thinned.dataset.batch)
}
}
\seealso{
\code{\link{env_thin}}, \code{\link{env_thin_b}}
}
