% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.f.data.prep.R
\name{load_thin_occ}
\alias{load_thin_occ}
\title{Load filtered occurrence data}
\usage{
load_thin_occ(thin.occ.l, from.disk = FALSE)
}
\arguments{
\item{thin.occ.l}{named list returned from "thin_b"}

\item{from.disk}{boolean. Read from disk or from one of thinned datasets stored in 'thin.occ.l' obj}
}
\value{
named list of thinned occurence data for each species in the list
}
\description{
Load filtered occurrence data from object returned by "thin_b" function
}
\examples{
\dontrun{
occ.locs <- load_thin_occ(thinned.dataset.batch)
}
}
\seealso{
\code{\link[spThin]{thin}}, \code{\link{thin_b}}
}
