% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{make.underscript}
\alias{make.underscript}
\title{Format (underscript) selected texts}
\usage{
make.underscript(x)
}
\arguments{
\item{x}{list of text to be formatted}
}
\value{
list of formatted text
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
plot(mxnt.mdls.preds.lst[[1]][[4]]) # MaxEnt predictions, based on the model selection criteria
}
