% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{make_underscript}
\alias{make_underscript}
\title{Plot compare climatic scenarios differences between model selection criteria}
\usage{
make_underscript(x)
}
\arguments{
\item{x}{list of text to be formatted}
}
\value{
list of formatted text
}
\description{
i.e. plot_scn_diff with panels for each of thresholds, and climatic scenario
 difference (i.e. current vs. climScn1) for each model being compared side
Format (underscript) selected texts for plotting
}
\details{
Format (underscript) selected texts (criteria used to select models) to be used on plotting.
}
\examples{
make_underscript(c("AUC (OR10p)", "AUC (ORlpt)", "OR10p (AUC)", "ORlpt (AUC)"))
}
\seealso{
\code{\link{plot_mdl_diff}}, \code{\link{plot_scn_diff_b}}
}
\keyword{internal}
