% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.2.f.mask_threshold.R
\name{mask_thr_projs_mscn_b}
\alias{mask_thr_projs_mscn_b}
\title{Remove areas from projections using polygons or binary rasters}
\usage{
mask_thr_projs_mscn_b(
  mtp.l,
  sp.mask,
  pred.args = "cloglog",
  append = T,
  mask.nm = "msk"
)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{sp.mask}{spatial object or list of spatial objects to use for masking suitability projections}

\item{pred.args}{arguments used for MaxEnt prediction. Can be found in mcmp$pred.args.
See resulting object from \code{\link{calib_mdl}} or \code{\link{calib_mdl_b}} functions}

\item{append}{Logical. Should append or not the resulting masked rasters?
Append (append=T): apply masks sequentially on the raster and save each one as separate layer.
Cascade (append=F): apply all masks on the raster and save as a single layer.}

\item{mask.nm}{Character vector. Names to identify the resulting masked rasters.
Should have the same length of 'sp.mask'.}
}
\description{
This function will remove (mask) areas from suitability projections using polygons or
binary rasters. This is usefull to filter predicted areas overlapping (e.g.) protected
areas or species habitats (forests, open areas, etc.).
}
\examples{
\dontrun{
mask_thr_projs_mscn_b(mtp.l=mods.thrshld.lst, sp.mask=NewWorld)
}
}
