% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.3.f.metrics.MOP.R
\name{mop}
\alias{mop}
\title{Extrapolation risk analysis}
\usage{
mop(M, G, p = 0.1, q = 0.1, min.M.sz = 100, filename = NULL,
  scn.nm = "", numCores = 1, ...)
}
\arguments{
\item{M}{RasterStack of environmental variables from calibration area}

\item{G}{RasterStack of environmental variables from projection area}

\item{p}{Percent of values, sampled from calibration area, used as
reference to calculate the MOP. Must be >0 and <=1.}

\item{q}{Quantile. Proportion of closest points in M is to be compared
with G to calculate the MOP. Must be >0 and <=1.}

\item{min.M.sz}{Threshold value to be used to compute OR}

\item{filename}{Output filename}

\item{scn.nm}{Name of climatic scenario to be looked for}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{...}{Additional arguments: 
  		
\code{datatype}{Character. Output data type (e.g. 'INT2S' or 'FLT4S'). 
See \code{\link{dataType}}. If no datatype is specified, 'FLT4S' is used,
unless this default value was changed with \code{\link{rasterOptions}}}
	
\code{overwrite}: Logical. If TRUE, "filename" will be overwritten if it exists
	
\code{progress}: Character. Set a value to show a progress bar. Valid values are "text" and "window".

\code{NAflag}: Numeric. To overwrite the default value used to represent \code{NA} in a file 
	
\code{bandorder}: Character. 'BIL', 'BIP', or 'BSQ'. For 'native' file formats only.
For some other formats you can use the 'options' argument (see below)
	
\code{options}: Character. File format specific GDAL options. E.g., when
writing a geotiff file you can use: \code{options=c("COMPRESS=NONE", "TFW=YES")}

You can use options=c("PROFILE=BASELINE") to create a plain  tif with no GeoTIFF tags. 
This can be useful when writing files to be read by applications intolerant of unrecognised tags. 
(see \url{http://www.gdal.org/frmt_gtiff.html})

NetCDF files have the following additional, optional, arguments: \code{varname}, \code{varunit}, \code{longname}, \code{xname}, \code{yname}, \code{zname}, \code{zunit}

\code{prj}: Logical. If \code{TRUE}, the crs is written to a .prj file. This can be useful 
when writing to an ascii file or another file type that does not store the crs
}
}
\value{
Data frame with average and variance of OR values across partition groups of data
}
\description{
This function will compute the omission rate (OR) for a species' AICc Averaged Model
from a 'mcmp' object, based on the selected threshold value.
}
\seealso{
\code{\link{mop_b}}
}
