% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.3.f.metrics.MOP.R
\name{mop_b}
\alias{mop_b}
\title{Extrapolation risk analysis for a list of species}
\usage{
mop_b(a.calib.l, a.proj.l, p = 0.1, q = 0.1, min.M.sz = 100,
  ref.scn = "current", format = "raster", numCores = 1)
}
\arguments{
\item{a.calib.l}{List of predictors (cropped environmental variables) for model tuning. Used in model calibration. Argument 'x' of dismo::maxent. Raster* object or SpatialGridDataFrame, containing grids with
predictor variables. These will be used to extract values from for the point locations. Can
also be a data.frame, in which case each column should be a predictor variable and each row
a presence or background record..}

\item{a.proj.l}{A list of Raster* objects or data.frames where models will be projected. Argument 'x' of dismo::predict}

\item{p}{Percent of values, sampled from calibration area, used as
reference to calculate the MOP. Must be >0 and <=1.}

\item{q}{Quantile. Proportion of closest points in M is to be compared
with G to calculate the MOP. Must be >0 and <=1.}

\item{min.M.sz}{Threshold value to be used to compute OR}

\item{ref.scn}{Selected climatic scenario to compare with all others. Usually "current" one.}

\item{format}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
Data frame with average and variance of OR values across partition groups of data
}
\description{
This function will compute the omission rate (OR) for each species' AICc Averaged Model
from a 'mcmp.l' object, based on the selected threshold value.
}
\seealso{
\code{\link{mop}}
}
