% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.3.f.metrics - MOP.R
\name{mopout}
\alias{mopout}
\title{Detecting values out of the environmental range of M}
\usage{
mopout(M, G)
}
\description{
Detecting values out of the environmental range of M
}
\keyword{internal}
