% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{mxnt.c.batch}
\alias{mxnt.c.batch}
\title{Calibrate MaxEnt models based on model selection criteria for several species}
\usage{
mxnt.c.batch(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will read a list of objects of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return selected maxent model calibrations and predictions. Each element on the list is usually a species.
a.proj.l, a.calib.l, occ.l are lists with occurence data, projection and calibration/predictor data.
Species in these lists must all be in the same order of species in ENMeval.o.
}
