% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{mxnt.c}
\alias{mxnt.c}
\title{Calibrate MaxEnt models based on model selection criteria}
\usage{
mxnt.c(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will read an object of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
calibrate the selected maxent models.
}
