% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.f.tuning.final.R
\name{mxnt.cp.batch}
\alias{mxnt.cp.batch}
\title{Calibrating and predicting selected models for several species}
\usage{
mxnt.cp.batch(ENMeval.res, a.calib.l, occ.l, formt = "raster",
  pred.args = c("outputformat=cloglog", "doclamp=true", "pictures=true"),
  wAICsum = 0.99, randomseed = F, responsecurves = T,
  arg1 = "noaddsamplestobackground", arg2 = "noautofeature", numCores = 1,
  parallelTunning = TRUE)
}
\arguments{
\item{ENMeval.res}{List of objects of class ENMevaluation}

\item{a.calib.l}{List of predictors (cropped environmental variables) for model tuning. Used in model calibration. Argument 'x' of dismo::maxent. Raster* object or SpatialGridDataFrame, containing grids with
predictor variables. These will be used to extract values from for the point locations. Can
also be a data.frame, in which case each column should be a predictor variable and each row
a presence or background record..}

\item{occ.l}{List of occurence data. See argument "occ" in mxnt.cp.}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{pred.args}{Charater. Argument 'args' of dismo::maxent. Additional argument that can be
passed to MaxEnt. See the MaxEnt help for more information. The R maxent function only uses the
arguments relevant to model fitting. There is no point in using args='outputformat=raw' when
*fitting* the model; but you can use arguments relevant for *prediction* when using the predict
function. Some other arguments do not apply at all to the R implementation. An example is
'outputfiletype', because the 'predict' function has its own 'filename' argument for that.}

\item{wAICsum}{cumulative sum of top ranked models for which arguments will be created}

\item{randomseed}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{responsecurves}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg1}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg2}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.
#' @examples
ENMeval.res.lst <- ENMevaluate.batch(occ.locs, occ.b.env, parallel = T , numCores = 7)
f.args(ENMeval.res.lst[[1]]@results)}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{parallelTunning}{Should parallelize within species (parallelTunning=TRUE) or between species (parallelTunning=FALSE)}
}
\value{
A list of objects returned from function "mxnt.cp"
}
\description{
This function will read a list of objects of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return selected maxent model calibrations and predictions. Each element on the list is usually a species.
a.proj.l, a.calib.l, occ.l are lists with occurence data, projection and calibration/predictor data.
Species in these lists must all be in the same order of species in ENMeval.res.
}
\examples{
mxnt.mdls.preds.lst <- mxnt.cp.batch(ENMeval.res=ENMeval.res.lst,
a.calib.l=occ.b.env, a.proj.l=areas.projection, occ.l=occ, wAICsum=0.99)
mxnt.mdls.preds.lst[[1]][[1]] # models [ENM]evaluated and selected using sum of wAICc
mxnt.mdls.preds.lst[[1]][[2]] # MaxEnt models
mxnt.mdls.preds.lst[[1]][[3]] # used prediction arguments
plot(mxnt.mdls.preds.lst[[1]][[4]]) # MaxEnt predictions, based on the model selection criteria
}
