% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.f.tuning.final.R
\name{mxnt.cpb}
\alias{mxnt.cpb}
\title{Calibrating and predicting selected models for several species}
\usage{
mxnt.cpb(ENMeval_res, a_calib.lst, a_proj.lst, occ_locs.lst, formt = "raster",
  pred.args = c("outputformat=cloglog", "doclamp=true", "pictures=true"),
  wAICsum = 0.99, randomseed = F, responsecurves = T,
  arg1 = "noaddsamplestobackground", arg2 = "noautofeature")
}
\arguments{
\item{ENMeval_res}{List of objects of class ENMevaluation}

\item{a_calib.lst}{List of predictor areas.}

\item{a_proj.lst}{List of projection areas. See argument "a_proj" in mxnt.cp.}

\item{occ_locs.lst}{List of occurence data. See argument "occ_locs" in mxnt.cp.}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{pred.args}{Charater. Argument 'args' of dismo::maxent. Additional argument that can be
passed to MaxEnt. See the MaxEnt help for more information. The R maxent function only uses the
arguments relevant to model fitting. There is no point in using args='outputformat=raw' when
*fitting* the model; but you can use arguments relevant for *prediction* when using the predict
function. Some other arguments do not apply at all to the R implementation. An example is
'outputfiletype', because the 'predict' function has its own 'filename' argument for that.}

\item{wAICsum}{cumulative sum of top ranked models for which arguments will be created}

\item{randomseed}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{responsecurves}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg1}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg2}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}
}
\value{
A list of objects returned from function "mxnt.cp"
}
\description{
This function will read a list of objects of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return selected maxent model calibrations and predictions. Each element on the list is usually a species.
a_proj.lst, a_calib.lst, occ_locs.lst are lists with occurence data, projection and calibration/predictor data.
Species in these lists must all be in the same order of species in ENMeval_res.
}
\examples{
mxnt.mdls.preds.lst <- mxnt.cpb(ENMeval_res=ENMeval_res.lst,
a_proj.lst=areas_projection, a_calib.lst=occ_b_env, occ_locs.lst=occ_locs, wAICsum=0.99)
mxnt.mdls.preds.lst[[1]][[1]] # models [ENM]evaluated and selected using sum of wAICc
mxnt.mdls.preds.lst[[1]][[2]] # MaxEnt models
mxnt.mdls.preds.lst[[1]][[3]] # used prediction arguments
plot(mxnt.mdls.preds.lst[[1]][[4]]) # MaxEnt predictions, based on the model selection criteria
}
