% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{mxnt.p.batch}
\alias{mxnt.p.batch}
\title{Project calibrated MaxEnt models}
\usage{
mxnt.p.batch(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will read an object returned by "calib_mdl", read the calibrated models and project into
new areas/climatic scenarios. These new projections will be returned together with (appended to)
the original object.
}
