% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.f.mscn.projection.R
\name{mxnt.p}
\alias{mxnt.p}
\title{Projecting Calibrated MaxEnt Models}
\usage{
mxnt.p(mcm, sp.nm, pred.nm = "fut", a.proj, formt = "raster",
  numCores = 1, parallelTunning = TRUE)
}
\arguments{
\item{mcm}{Objects returned by "mxnt.c", containing calibrated models.}

\item{sp.nm}{Species name. Used to name the output folder}

\item{pred.nm}{Character. Prefix to add to projection name (e.g. "fut" or "past")}

\item{a.proj}{A Raster* object or a data.frame where models will be projected. Argument 'x' of dismo::predict}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{parallelTunning}{Should parallelize within species (parallelTunning=TRUE) or between species (parallelTunning=FALSE)}
}
\value{
A list containing all the items returned from function "mxnt.c", plus the projection specified in a.proj.
Each projection is a raster stack containing model projections ('mxnt.preds'), where each layer is a projection based on
a specific model selection criteria (i.e. AvgAICc, LowAICc, Mean.ORmin, Mean.OR10, Mean.AUCmin, Mean.AUC10)
}
\description{
This function will read an object returned by "mxnt.c", read the calibrated models and project into
new areas/climatic scenarios. These new projections will be returned together with (appended to)
the original object.
}
\seealso{
\code{\link{f.args}}, \code{\link{mxnt.c}}, \code{\link{mxnt.c.batch}}, \code{\link[dismo]{maxent}},
\code{\link[ENMeval]{ENMevaluate}}, \code{\link{mxnt.p.batch.mscn}}
}
\keyword{internal}

