% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.f.mscn.projection.R
\name{mxnt.pb.mscn}
\alias{mxnt.pb.mscn}
\title{Projecting calibrated MaxEnt models for several species onto multiple environmental scenarios}
\usage{
mxnt.pb.mscn(mxnt.c.mdls.lst, a_proj.lst, formt = "raster")
}
\arguments{
\item{mxnt.c.mdls.lst}{A list of objects returned by "mxnt.cp", containing calibrated models.}

\item{a_proj.lst}{List of projection areas. See argument "a_proj" in mxnt.cp.}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}
}
\value{
A list of objects returned from function "mxnt.p", containing the new (multiple) projections for each element (species) of the list
}
\description{
This function will read an object returned by "mxnt.cpb", read the calibrated models and project into
several environmental (areas/climatic) scenarios (specified in a_proj.lst). These new projections will be returned together with (appended to)
each element (species) the original object.
}
\examples{
mxnt.mdls.preds.pf <- mxnt.pb.Mscn(mxnt.mdls.preds.lst, a_proj.lst = area_projection.pf)
}
