% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.f.mscn.projection.R
\name{mxnt.pb}
\alias{mxnt.pb}
\title{Projecting calibrated MaxEnt models for several species}
\usage{
mxnt.pb(mxnt.c.mdls.lst, pred.nm = "fut", a_proj.lst, formt = "raster")
}
\arguments{
\item{mxnt.c.mdls.lst}{A list of objects returned by "mxnt.cp", containing calibrated models.}

\item{pred.nm}{Character. Prefix to add to projection name (e.g. "fut" or "past")}

\item{a_proj.lst}{List of projection areas. See argument "a_proj" in mxnt.cp.}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}
}
\value{
A list of objects returned from function "mxnt.p"
}
\description{
This function will read an object returned by "mxnt.cpb", read the calibrated models and project into
new areas/climatic scenarios. These new projections will be returned together with (appended to)
each element (species) the original object
}
\examples{
mxnt.mdls.preds <- mxnt.pb(mxnt.c.mdls.lst = mxnt.mdls.preds.lst[1],
}
