% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.1.accessory.functions.R
\name{north_arrow}
\alias{north_arrow}
\title{Draw north arrow in plotted map}
\usage{
north_arrow(
  xb,
  yb,
  len,
  lab = "NORTH",
  cex.lab = 1,
  border = "black",
  fill = c("white", border),
  ...
)
}
\arguments{
\item{xb}{Position at x axis}

\item{yb}{Position at y axis}

\item{len}{Arrow length}

\item{lab}{Label}

\item{cex.lab}{A numerical vector giving the amount by which plotting
label characters should be scaled relative to the default.}

\item{border}{the color to draw the border.  The default, \code{NULL},
    means to use \code{\link[graphics]{par}("fg")}.  Use \code{border = NA} to
    omit borders.

    For compatibility with S, \code{border} can also be logical, in
    which case \code{FALSE} is equivalent to \code{NA} (borders omitted)
    and \code{TRUE} is equivalent to \code{NULL} (use the foreground colour),
  }

\item{fill}{The color for filling polygons of both sides of the north arrow.
The default, c("white", border), fills one side with white and the other side
with the color of the border. Two colors should be provided.}

\item{...}{graphical parameters such as \code{xpd}, \code{lend},
    \code{ljoin} and \code{lmitre} can be given as arguments.}
}
\description{
This function will draw a simple north arrow in a plotted map
}
\author{
Neander M. Heming
}
