% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{plot_mdl_diff}
\alias{plot_mdl_diff}
\title{Plot differences in suitable areas between models selected using distinct criteria}
\usage{
plot_mdl_diff(
  mcmp,
  mtp,
  sp.nm = "species",
  basemap = NULL,
  save = FALSE,
  mfrow = c(0, 0),
  numCores = 1,
  msnm = c("AUC10", "AUCmtp", "OR10", "ORmtp"),
  msr = c("AUC (OR10p)", "AUC (ORlpt)", "OR10p (AUC)", "ORlpt (AUC)"),
  scnm = c("cc26bi70", "cc45bi70", "cc60bi70", "cc85bi70", "mp26bi70", "mp45bi70",
    "mp85bi70", "mr26bi70", "mr45bi70", "mr60bi50", "mr85bi70", "cclgmbi", "ccmidbi",
    "lig_30s_bio_", "melgmbi", "memidbi", "mrlgmbi", "mrmidbi", "ncurrent", "current"),
  scr = c("2070-CCSM4-rcp2.6", "2070-CCSM4-rcp4.5", "2070-CCSM4-rcp6.0",
    "2070-CCSM4-rcp8.5", "2070-MPI-ESM-LR-rcp2.6", "2070-MPI-ESM-LR-rcp4.5",
    "2070-MPI-ESM-LR-rcp8.5", "2070-MIROC-ESM-rcp2.6", "2070-MIROC-ESM-rcp4.5",
    "2070-MIROC-ESM-rcp6.0", "2070-MIROC-ESM-rcp8.5", "LGM-CCSM4", "MH-CCSM4",
    "LIG-CCSM3", "LGM-MPI-ESM-P", "MH-MPI-ESM-P", "LGM-MIROC-ESM", "MH-MIROC-ESM",
    "Near Current", "Near Current")
)
}
\arguments{
\item{mcmp}{Species "i" of a object returned by "proj_mdl_b", containing a list of
calibrated models and model projections for each species}

\item{mtp}{Stack or brick of thresholded predictions}

\item{sp.nm}{Species name. Used to name the output folder}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}

\item{save}{Logical. If TRUE will save plots in pdf.}

\item{mfrow}{Number of rows and cols to draw plot (see ?par for details). If}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{msnm}{Character vector. Short names of model selection criteria to be replaced. Same as given in
Model names (ex. AUC10, AUCmtp, OR10, ORmtp)}

\item{msr}{Character vector. Long names of model selection criteria to replace the short names. Must
be in same order of argument 'msnm'}

\item{scnm}{Character vector. Short names of climatic scenarios to be replaced. Ex. "cc26bi70", "cc45bi70"}

\item{scr}{Character vector. Long names of climatic scenarios to replace the short names. Must
be in same order of argument 'scnm'. Ex. "2070-CCSM4-rcp2.6", "2070-CCSM4-rcp4.5"}
}
\description{
Will plot (or save as PDF) differences between predictions of models selected
 using distinct model selection criteria (for multiple climatic scenarios). PDFs will be saved
 within the folder 'Mdls.thrshld/figs'.
}
\details{
Plot differences between predictions of models selected using distinct
criteria (e.g. "AvgAIC", "LowAIC", "OR", "AUC") for multiple climatic scenarios
}
\examples{
\dontrun{
plot_mdl_diff(mcmp=mxnt.mdls.preds.lst[[1]], mtp.l=mods.thrshld.lst[[1]], basemap=NewWorld)
}
}
\seealso{
\code{\link{plot_scn_diff_b}}
}
