% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{plot_scn_diff_b}
\alias{plot_scn_diff_b}
\title{Plot differences in suitable areas between climatic scenarios for multiple species}
\usage{
plot_scn_diff_b(mcmp.l, mtp.l, mSel = mcmp.l[[1]]$mSel,
  ref.scn = "current", basemap = NULL, save = FALSE, numCores = 1,
  msnm = c("AUC10", "AUCmtp", "OR10", "ORmtp"), msr = c("AUC (OR10p)",
  "AUC (ORlpt)", "OR10p (AUC)", "ORlpt (AUC)"), scnm = c("cc26bi70",
  "cc45bi70", "cc60bi70", "cc85bi70", "mp26bi70", "mp45bi70", "mp85bi70",
  "mr26bi70", "mr45bi70", "mr60bi50", "mr85bi70", "cclgmbi", "ccmidbi",
  "lig_30s_bio_", "melgmbi", "memidbi", "mrlgmbi", "mrmidbi",
  "mxnt.preds"), scr = c("2070-CCSM4-rcp2.6", "2070-CCSM4-rcp4.5",
  "2070-CCSM4-rcp6.0", "2070-CCSM4-rcp8.5", "2070-MPI-ESM-LR-rcp2.6",
  "2070-MPI-ESM-LR-rcp4.5", "2070-MPI-ESM-LR-rcp8.5",
  "2070-MIROC-ESM-rcp2.6", "2070-MIROC-ESM-rcp4.5",
  "2070-MIROC-ESM-rcp6.0", "2070-MIROC-ESM-rcp8.5", "LGM-CCSM4",
  "MH-CCSM4", "LIG-CCSM3", "LGM-MPI-ESM-P", "MH-MPI-ESM-P",
  "LGM-MIROC-ESM", "MH-MIROC-ESM", "Present"))
}
\arguments{
\item{mcmp.l}{Object returned by "proj_mdl_b", containing a list of calibrated models
and model projections for each species.}

\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{mSel}{Name of selection criteria to be compared: AvgAIC, LowAIC, avg.test.AUC10pct, avg.test.AUC.MTP,
avg.test.or10pct, avg.test.orMTP}

\item{ref.scn}{Selected climatic scenario to compare with all others. Usually "current" one.}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}

\item{save}{Logical. If TRUE will save plots in pdf.}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{msnm}{Character vector. Short names of model selection criteria to be replaced. Same as given in
Model names (ex. AUC10, AUCmtp, OR10, ORmtp)}

\item{msr}{Character vector. Long names of model selection criteria to replace the short names. Must
be in same order of argument 'msnm'}

\item{scnm}{Character vector. Short names of climatic scenarios to be replaced. Ex. "cc26bi70", "cc45bi70"}

\item{scr}{Character vector. Long names of climatic scenarios to replace the short names. Must
be in same order of argument 'scnm'. Ex. "2070-CCSM4-rcp2.6", "2070-CCSM4-rcp4.5"}
}
\value{
won't return any object. Will save pdf's with differences among model predictions (for multiple climatic scenarios)
}
\description{
Plot differences between a selected climatic scenario and all other climatic scenarios for each species.
This function will plota and (optionally) save the figures on pdf files in the folder "Mdls.thrshld/figs".
}
\examples{
\dontrun{
plot_scn_diff_b(mcmp.l=mxnt.mdls.preds.lst, mtp.l=mods.thrshld.lst)
}
}
\seealso{
\code{\link{plot_scn_diff}}
}
