% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{plotMdlDiff}
\alias{plotMdlDiff}
\title{Plot differences in suitable areas between models selected using distinct criteria for multiple species}
\usage{
plotMdlDiff(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Plot differences between predictions of models selected using distinct
criteria (e.g. "AvgAIC", "LowAIC", "OR", "AUC") for multiple climatic scenarios and multiple species
}
