% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{plotScnDiff}
\alias{plotScnDiff}
\title{Plot differences in suitable areas between models selected using distinct criteria}
\usage{
plotScnDiff(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Plot differences between predictions of models selected using distinct
criteria (e.g. "AvgAIC", "LowAIC", "OR", "AUC") for multiple climatic scenarios
}
\details{
Plot differences in suitable areas between climatic scenarios for multiple species

Plot differences between a selected climatic scenario and all other climatic scenarios for each species.
This function will plota and (optionally) save the figures on pdf files in the folder "Mdls.thrshld/figs".
}
