% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{poly.c.batch}
\alias{poly.c.batch}
\title{Create occ polygon for several species}
\usage{
poly.c.batch(spp.occ.list, k = 1, c.m = "AP", r = 2, q = 0.3,
  distance = "euclidean", min.nc = 2, max.nc = 20, method = "mcquitty",
  index = "all", alphaBeale = 0.1, convex = T, alpha = 10, plot = T,
  save = T, save.pts = F, numCores = 1,
  crs.set = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{spp.occ.list}{A named list of species occurence points, either as "data.frame" or "SpatialPoints"/"SpatialPointsDataFrame"}

\item{k}{number of polygons to create based on coordinates}

\item{c.m}{clustering method to find the best number of clusters (k). Currently E (Elbow) or (Affinity Propagation).}

\item{convex}{Logical. Convex (T) or concave (F) polygon}

\item{alpha}{Value of \eqn{\alpha}.}

\item{plot}{logical. Plot results or not?}

\item{save}{Should save polygons on disk?}

\item{save.pts}{logical. Save each species' occurence points as shapefile?}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\description{
This function will use a list of coordinates of species occurence data and create minimum convex polygons
for each element in the list.
It is possible to create concave or convex polygons, create several small polygons based on clusters
of points.
}
\examples{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- poly.c.batch(spp.occ.list)
occ.polys <- poly.c.batch(spp.occ.list, convex=TRUE, alpha=10)
}
\seealso{
\code{\link{poly.c}}, \code{\link{poly.splt}}, \code{\link[NbClust]{NbClust}}
#' @return A named list of spatial polygons built using coordinates
}

