% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{poly.c}
\alias{poly.c}
\title{Create polygon based on species occurence coordinates}
\usage{
poly.c(occ.spdf, sp.nm = "sp.nm", convex = T, alpha = 10, save = T,
  crs.set = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{occ.spdf}{A spatial data.frame of coordinates, usually species occurence coordinates}

\item{sp.nm}{Species name, used on saving shapefile}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{Value of \eqn{\alpha}.}

\item{save}{Should save polygons on disk?}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\value{
spatial polygon of occurencies built using coordinates
}
\description{
Create polygon based on species occurence coordinates
}
\examples{
occ.poly <- poly.c(occ.spdf, sp.nm="occ.poly")
plot(occ.poly)
}
