% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{poly.splt}
\alias{poly.splt}
\title{split a species occ polygon (whenever distribution seems disjoint) into K polygons and save in a single .shp}
\usage{
poly.splt(spp.occ, k = 2, convex = T, alpha = 10, sp.nm = "sp1",
  crs.set = NULL)
}
\arguments{
\item{spp.occ}{species occurence coordinates}

\item{k}{number of polygons to create based on coordinates}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{Value of \eqn{\alpha}.}

\item{sp.nm}{name used to save file}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\value{
spatial polygons built using coordinates
}
\description{
split a species occ polygon (whenever distribution seems disjoint) into K polygons and save in a single .shp
}
\examples{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- poly.c.batch(spp.occ.list)
occ.polys$Bvarieg <- poly.splt(spp.occ = spp.occ.list$Bvarieg, k=5)
}
