% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{polyC}
\alias{polyC}
\title{Create minimum convex polygon based on species occurence data}
\usage{
polyC(occ.spdf, sp.nm = "species", convex = TRUE, alpha = 10,
  save = TRUE, crs.set = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{occ.spdf}{An object of class SpatialPoints or SpatialPointsDataFrame (e.g. species occurence coordinates)}

\item{sp.nm}{Species name, used on saving shapefile}

\item{convex}{Logical. Convex (T) or concave (F) polygon}

\item{alpha}{Value of \eqn{\alpha}.}

\item{save}{Should save polygons on disk?}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\value{
An object of class SpatialPolygons or SpatialPolygonsDataFrame. Polygon built using coordinates of species occurence data
}
\description{
This function will create minimum convex polygon based on coordinates of species occurence data.
}
\examples{
occ.poly <- polyC(occ.spdf, sp.nm="occ.poly")
plot(occ.poly)
}
\seealso{
\code{\link{polyCB}}, \code{\link{polySplt}}
}
