% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.a.batch}
\alias{pred.a.batch}
\title{Cut area for projection based on a list of SpatialPolygons}
\usage{
pred.a.batch(pred.polys, env.uncut, prj.nm = "")
}
\arguments{
\item{pred.polys}{list of SpatialPolygons (usually of based on species occ points)

#' @seealso \code{\link{pred.a.poly}}, \code{\link{pred.a.poly.batch}},
\code{\link{pred.a}}, \code{\link{pred.a.batch.mscn}},
\code{\link{pred.a.rst}}, \code{\link{pred.a.batch.rst}}, \code{\link{pred.a.batch.rst.mscn}}}

\item{env.uncut}{raster/brick of environmental variables to be cut}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}
}
\value{
named list of cropped raster or brick
}
\description{
This function is a wrapper for "pred.a". See ?pred.a
It works with a named list of pred.poly to delimitate the projection area for each of the species.
}
\examples{
area.projection <- pred.a.batch(pred.polys, env.uncut)
plot(area.projection[[1]][[1]])
}
