% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{pred.a.batch.rst.mscn}
\alias{pred.a.batch.rst.mscn}
\title{Cut multiple projection areas of multiple species based on a single SpatialPolygon object (e.g. Ecoregion)}
\usage{
pred.a.batch.rst.mscn(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will use a single SpatialPolygon to crop/mask multiple raster/brick objects
to be used on model projections.
}
