% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{pred.a.poly.batch}
\alias{pred.a.poly.batch}
\title{Select area for projection based on the extent of occ points for multiple species}
\usage{
pred.a.poly.batch(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function is a wrapper for "cut_projarea". See ?cut_projarea
It works with a named list of occ.polys to delimit the projection area for each of the species.
}
