% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.a.polyb}
\alias{pred.a.polyb}
\title{Select area for projection based on the extent of occ points}
\usage{
pred.a.polyb(occ_polys, deg.incr = NULL, mult = 1, buffer = F, same = T)
}
\arguments{
\item{occ_polys}{list of occ_poly SpatialPolygons.  see ?pred.a.poly for details}

\item{deg.incr}{used to manually set the increase in prediction area, relative to occ_poly}

\item{mult}{how much increase or decrease buffer}

\item{buffer}{should the area be cut using a buffer around occ_poly?}

\item{same}{should latitudinal and longitudinal increase vary independently?}
}
\value{
named list of SpatialPolygons (enlarged occ_poly)
}
\description{
This function is a wrapper for "pred.a". See ?pred.a
It works with a named list of occ_polys to delimitate the projection area for each of the species
}
\examples{
area_projection <- pred.a.polyb(occ_polys, mult=.55, buffer=F)
plot(area_projection[[1]][[1]])
}
