% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.a}
\alias{pred.a}
\title{Cut area for projection based on SpatialPolygons}
\usage{
pred.a(pred.poly, env.uncut, prj.nm = "", sp.nm = "sp")
}
\arguments{
\item{pred.poly}{list of SpatialPolygons (usually of based on species occ points)}

\item{env.uncut}{raster/brick of environmental variables to be cut}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{sp.nm}{name (of species) to give to saved object}
}
\value{
raster or brick cropped based on SpatialPolygons
}
\description{
This function will use SpatialPolygons to crop/mask raster/brick objects to be used on model projections.
}
