% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{pred.a.rst}
\alias{pred.a.rst}
\title{Cut a projection area based on a SpatialPolygon (e.g. Ecoregion)}
\usage{
pred.a.rst(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will use a single SpatialPolygon to crop/mask raster/brick objects to be used on model projections.
}
