% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.ab.mscn}
\alias{pred.ab.mscn}
\title{Cut multiple projection areas (climatic scenarios) for multiple species (list of SpatialPolygons)}
\usage{
pred.ab.mscn(pred_polys, env_uncut.l, prj.nm = "", cores = 1)
}
\arguments{
\item{pred_polys}{list of SpatialPolygons (usually of based on species occ points)}

\item{env_uncut.l}{list of raster/brick of environmental variables to be cut}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{cores}{specify number of cores if aim run in parallel}
}
\value{
named list of cropped list of raster/brick of environmental variables
}
\description{
This function is a wrapper for "pred.a". See ?pred.a. This function delimitates the projection area for
 each of the species contained in the pred_polys named list and crops
 multiple rasters/bricks (i.e. representing distinct climatic scenaries) based on the same criteria for each species
}
