% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.ab.rst.mscn}
\alias{pred.ab.rst.mscn}
\title{Cut multiple projection areas of multiple species based on a unique SpatialPolygon (e.g. Ecoregion)}
\usage{
pred.ab.rst.mscn(area.p, env.uncut.l, occ.polys, mask = F, prj.nm = "",
  sp.nm = "a.proj4mult.spp", cores = 1)
}
\arguments{
\item{area.p}{SpatialPolygon to be used as reference to crop/mask environmental variables of all species}

\item{occ.polys}{list of occ.poly SpatialPolygons.  see ?pred.a.poly for details}

\item{mask}{Should mask raster? (i.e. only use area inside polygon. See ?raster::mask for details) or use all spatial extent of area.p}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{sp.nm}{name (of species) to give to saved object}
}
\value{
list of list with multiple environmental layers (raster/brick) cutted
}
\description{
This function will use a single SpatialPolygon to crop/mask multiple raster/brick objects
to be used on model projections.
}
