% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.ab.rst}
\alias{pred.ab.rst}
\title{Cut projection areas of multiple species based on a unique SpatialPolygon (e.g. Ecoregion)}
\usage{
pred.ab.rst(area_p, env_uncut, occ_polys, mask = F, prj.nm = "",
  sp.nm = "a.proj4mult.spp")
}
\arguments{
\item{area_p}{SpatialPolygon to be used as reference to crop/mask environmental variables of all species}

\item{env_uncut}{raster/brick of environmental variables to be cut}

\item{occ_polys}{list of occ_poly SpatialPolygons.  see ?pred.a.poly for details}

\item{mask}{Should mask raster? (i.e. only use area inside polygon. See ?raster::mask for details) or use all spatial extent of area_p}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{sp.nm}{name (of species) to give to saved object}
}
\value{
list of environmental layers (raster/brick) cutted
}
\description{
This function will use a single SpatialPolygon to crop/mask raster/brick objects to be used on model projections.
}
