% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{predArB}
\alias{predArB}
\title{Cut area for projection based on a list of SpatialPolygons}
\usage{
predArB(pred.polys, env.uncut, prj.nm = "")
}
\arguments{
\item{pred.polys}{list of SpatialPolygons (usually of based on species occ points)

#' @seealso \code{\link{predArPoly}}, \code{\link{predArPolyB}},
\code{\link{predAr}}, \code{\link{predArMscnB}},
\code{\link{predArRst}}, \code{\link{predArRstB}}, \code{\link{predArRstMscnB}}}

\item{env.uncut}{raster/brick of environmental variables to be cut}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}
}
\value{
named list of cropped raster or brick
}
\description{
This function is a wrapper for "predAr". See ?predAr
It works with a named list of pred.poly to delimit the projection area for each of the species.
}
\examples{
area.projection <- predArB(pred.polys, env.uncut)
plot(area.projection[[1]][[1]])
}
