% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{predArRstB}
\alias{predArRstB}
\title{Cut projection areas of multiple species based on a single SpatialPolygon object (e.g. Ecoregion)}
\usage{
predArRstB(area.p, env.uncut, occ.polys, mask = FALSE, prj.nm = "",
  sp.nm = "mult.spp")
}
\arguments{
\item{area.p}{SpatialPolygon to be used as reference to crop/mask environmental variables of all species}

\item{env.uncut}{raster/brick of environmental variables to be cut}

\item{occ.polys}{list of occ.poly SpatialPolygons.  see ?predArPoly for details.}

\item{mask}{Should mask raster? (i.e. only use area inside polygon. See ?raster::mask for details) or
use all spatial extent of area.p

#' @seealso \code{\link{predArPoly}}, \code{\link{predArPolyB}},
\code{\link{predAr}}, \code{\link{predArB}}, \code{\link{predArMscnB}},
\code{\link{predArRst}}, \code{\link{predArRstMscnB}}}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{sp.nm}{name (of species) to give to saved object}
}
\value{
list of environmental layers (raster/brick) cutted
}
\description{
This function will use a single SpatialPolygon to crop/mask raster/brick objects to be used on model projections.
}
