% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1.f.mscn.projection.R
\name{proj_mdl}
\alias{proj_mdl}
\title{Project calibrated MaxEnt models}
\usage{
proj_mdl(
  mcm,
  sp.nm = "species",
  pred.nm = "fut",
  a.proj,
  format = "raster",
  numCores = 1,
  parallelTunning = TRUE
)
}
\arguments{
\item{mcm}{Objects returned by "calib_mdl", containing calibrated models.}

\item{sp.nm}{Species name. Used to name the output folder}

\item{pred.nm}{Character. Prefix to add to projection name (e.g. "fut" or "past")}

\item{a.proj}{A Raster* object or a data.frame where models will be projected. Argument 'x' of dismo::predict}

\item{format}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{parallelTunning}{Should parallelize within species (parallelTunning=TRUE) or between species (parallelTunning=FALSE)}
}
\value{
A list containing all the items returned from function "calib_mdl", plus the projection specified in a.proj.
Each projection is a raster stack containing model projections ('mxnt.preds'), where each layer is a projection based on
a specific model selection criteria (i.e. AvgAIC, LowAIC, avg.test.orMTP, avg.test.or10pct, avg.test.AUC.MTP, avg.test.AUC10pct)
}
\description{
This function will read an object returned by "calib_mdl", read the calibrated models and project into
new areas/climatic scenarios. These new projections will be returned together with (appended to)
the original object.
}
\seealso{
\code{\link{mod_sel}}, \code{\link{calib_mdl}}, \code{\link{calib_mdl_b}}, \code{\link[dismo]{maxent}},
\code{\link[ENMeval]{ENMevaluate}}, \code{\link{ENMevaluate_b}}, \code{\link{proj_mdl_b}}
}
