% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.f.mscn.projection.R
\name{proj_mdl_b}
\alias{proj_mdl_b}
\title{Project calibrated MaxEnt models for several species onto multiple environmental scenarios}
\usage{
proj_mdl_b(mcm.l, a.proj.l, format = "raster", numCores = 1,
  parallelTunning = TRUE)
}
\arguments{
\item{mcm.l}{A list of objects returned by "calib_mdl", containing calibrated models.}

\item{a.proj.l}{A list of Raster* objects or data.frames where models will be projected. Argument 'x' of dismo::predict}

\item{format}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{parallelTunning}{Should parallelize within species (parallelTunning=TRUE) or between species (parallelTunning=FALSE)}
}
\value{
A 'mcmp.spl' object. A list of objects returned from function "proj_mdl", containing the new (multiple) projections for each element (species) of the list
}
\description{
This function will read an object returned by "calib_mdl_b", read the calibrated models and project into
several environmental (areas/climatic) scenarios (specified in a.proj.l). These new projections will be returned together with (appended to)
each element (species) the original object.
}
\examples{
\dontrun{
mxnt.mdls.preds.pf <- proj_mdl_b(mcm.l = mxnt.mdls.preds.lst, a.proj.l = area.projection.pf)
}
}
\seealso{
\code{\link{mod_sel}}, \code{\link{calib_mdl}}, \code{\link{calib_mdl_b}}, \code{\link[dismo]{maxent}},
\code{\link[ENMeval]{ENMevaluate}}, \code{\link{ENMevaluate_b}}, \code{\link{proj_mdl}}
}
