% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics-area_gain_loss.R
\name{range_shift}
\alias{range_shift}
\title{Map shifts in suitable areas}
\usage{
range_shift(
  mtp,
  ref.scn = "ncurrent",
  sp.nm = "species",
  numCores = 1,
  format = "raster"
)
}
\arguments{
\item{mtp}{Stack or brick of thresholded predictions}

\item{ref.scn}{Selected climatic scenario to compare with all others. Usually "ncurrent".
any (row or col) is 0, it will draw the layout automatically.}

\item{sp.nm}{Species name. Used to name the output folder}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{format}{Character. Output file type. Argument 'format' of raster::writeRaster}
}
\value{
A list containing rasters of differences between predictions of climatic scenarios
}
\description{
Map shifts in suitable areas between a selected climatic scenario
and all other climatic scenarios.
}
\examples{
\dontrun{
range_shift(mcmp.l=mxnt.mdls.preds.lst, mtp.l=mods.thrshld.lst)
}
}
\seealso{
\code{\link{range_shift_b}}
}
