% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1.f.metrics-area_gain_loss.R
\name{range_shift_b}
\alias{range_shift_b}
\title{Map shifts in suitable areas}
\usage{
range_shift_b(mtp.l, ref.scn = "ncurrent", numCores = 1, format = "raster")
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{ref.scn}{Selected climatic scenario to compare with all others. Usually "ncurrent".
any (row or col) is 0, it will draw the layout automatically.}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{format}{Character. Output file type. Argument 'format' of raster::writeRaster}
}
\value{
Return a nested list, containig rasters with differences between
a selected climatic scenario and all other climatic scenarios.
}
\description{
Map shifts in suitable areas between a selected climatic scenario
and all other climatic scenarios for each species.
}
\examples{
\dontrun{
spp_diff <- range_shift_b(mods.thrshld.lst, ref.scn = "ncurrent")

col <- c("red", "gray", "blue") #RColorBrewer::brewer.pal(9,"YlOrRd")
breaks <- round(seq(from=-1, to=1, .666), 2)
colors <- colorRampPalette(col)(length(breaks)-1)

plot(spp_diff[[1]][[1]][[1]], col=colors, breaks=breaks)
}
}
\seealso{
\code{\link{range_shift}}
}
