% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1b.f.occ.thin.R
\name{rescale}
\alias{rescale}
\title{Re-scale a variable to a desired range.}
\usage{
rescale(x, x.min = NULL, x.max = NULL, new.min = 0, new.max = 1)
}
\arguments{
\item{x}{vector of values}

\item{x.min}{minimum values to be considered}

\item{x.max}{maximu values to be considered}

\item{new.min}{new minimum values}

\item{new.max}{new maximum values}
}
\description{
Re-scale a variable to a desired range.
}
