% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{sel_env}
\alias{sel_env}
\title{Find, optionally remove, highly correlated variables from a raster brick/stack}
\usage{
sel_env(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Find, optionally remove, highly correlated variables from a raster brick/stack
}
