% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{sel_env_b}
\alias{sel_env_b}
\title{Find, optionally remove, highly correlated variables from a list (several species) of raster brick/stack}
\usage{
sel_env_b(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Find, optionally remove, highly correlated variables from a list (several species) of raster brick/stack
}
