% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2b.f.sel.env.R
\name{select_vars_b}
\alias{select_vars_b}
\title{Find, optionally remove, highly correlated variables from a list (several species) of raster brick/stack}
\usage{
select_vars_b(
  env.l,
  cutoff = 0.9,
  corr.mat.l = NULL,
  sample.prop = 0.1,
  names.only = F,
  plot.dend = T,
  rm.old = F,
  filename = NULL
)
}
\arguments{
\item{env.l}{List of raster brick/stack.}

\item{cutoff}{A numeric value for the pair-wise absolute correlation cutoff}

\item{corr.mat.l}{List of correlation matrices from which variables will be selected. If the correlation
matrix was already computed from env, you can just input here and choose other cutoff values for
selecting variable layers.}

\item{names.only}{Logical. Return only the names of selected variables (T) or return
the raster brick containing the least correlated variables (F).}

\item{plot.dend}{Logical. Plot dendrogram of correlation distance between variables,
showing cutoff limit and selected (black) and discarded (red) variables}

\item{rm.old}{Logical. Remove (T) old env variables from folder?}

\item{filename}{Output filename}
}
\description{
This function is a wrapper for ENMwizard::env.sel. See ?ENMwizard::env.sel for details
It works with a named list of environmental variables (raster brick/stack).
It creates a correlation matrix for the layers of a raster brick/stack for each species
(item in the list)
and returns a brick containing the least correlated variables below the cutoff value.
}
\examples{
\dontrun{
select_vars_b(occ.b.env, .9, names.only=T)
occ.b.env <- select_vars_b(occ.b.env, .9, names.only=F, rm.old=F)
}
}
\seealso{
\code{\link{select_vars}}, \code{\link[caret]{findCorrelation}}
}
