% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{set_projarea_b}
\alias{set_projarea_b}
\title{Select area for projection based on the extent of occ points for multiple species}
\usage{
set_projarea_b(
  occ.polys,
  deg.incr = NULL,
  mult = 1,
  buffer = FALSE,
  same = TRUE
)
}
\arguments{
\item{occ.polys}{list of occ.poly SpatialPolygons.  see ?set_projarea for details.}

\item{deg.incr}{used to manually set the increase in prediction area, relative to occ.poly}

\item{mult}{how much increase or decrease buffer}

\item{buffer}{should the area be cut using a buffer around occ.poly?}

\item{same}{should latitudinal and longitudinal increase vary independently?}
}
\value{
named list of SpatialPolygons (enlarged occ.poly)
}
\description{
This function is a wrapper for "cut_projarea". See ?cut_projarea
It works with a named list of occ.polys to delimit the projection area for each of the species.
}
\examples{
\dontrun{
area.projection <- set_projarea_b(occ.polys, mult=.55, buffer=FALSE)
plot(area.projection[[1]][[1]])
}
}
\seealso{
\code{\link{set_projarea}}
}
