% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{thin.batch}
\alias{thin.batch}
\title{Filter each species' occurence dataset}
\usage{
thin.batch(loc.data.lst, lat.col = "LAT", long.col = "LONG",
  spec.col = "SPEC", thin.par = 10, reps = 10,
  locs.thinned.list.return = TRUE, write.files = TRUE, max.files = 1,
  write.log.file = TRUE)
}
\arguments{
\item{loc.data.lst}{Named list containing species occ data}

\item{lat.col}{See ?thin of spThin package}

\item{long.col}{See ?thin of spThin package}

\item{spec.col}{See ?thin of spThin package}

\item{thin.par}{See ?thin of spThin package}

\item{reps}{See ?thin of spThin package}

\item{locs.thinned.list.return}{See ?thin of spThin package}

\item{write.files}{See ?thin of spThin package}

\item{max.files}{See ?thin of spThin package}

\item{write.log.file}{See ?thin of spThin package}
}
\value{
Named list containing thinned datasets for each species. See ?thin of spThin package. Also, by default it saves log file and the first thinned dataset in the folder "occ_thinned_full".
}
\description{
Filter each species' occurence dataset
}
\examples{
thinned_dataset_batch <- thin.batch(loc.data.lst = spp.occ.list)
plotThin(thinned_dataset_batch[[1]])
length(thinned_dataset_batch[[1]])
}
