% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.f.threshold.R
\name{thr}
\alias{thr}
\title{Apply threshold for MaxEnt projections of a species}
\usage{
thr(mcmp, scn.nm = "", thrshld.i = 4:6, path.mdls = NULL,
  sp.nm = "species")
}
\arguments{
\item{mcmp}{Species "i" of a object returned by "mxntProjB", containing a list of
calibrated models and model projections for each species}

\item{scn.nm}{Name of climatic scenario to be looked for}

\item{thrshld.i}{List of threshold criteria to be applied. Use numbers to choose the desired one(s). Current options:
1. Fixed.cumulative.value.1 (fcv1);
2. Fixed.cumulative.value.5 (fcv5);
3. Fixed.cumulative.value.10 (fcv10);
4. Minimum.training.presence (mtp);
5. 10.percentile.training.presence (x10ptp);
6. Equal.training.sensitivity.and.specificity (etss);
7. Maximum.training.sensitivity.plus.specificity (mtss);
8. Balance.training.omission.predicted.area.and.threshold.value (bto);
9. Equate.entropy.of.thresholded.and.original.distributions (eetd).}

\item{path.mdls}{Path where thresholded rasters will be saved}

\item{sp.nm}{Species name. Used to name the output folder}
}
\value{
Stack or brick of thresholded predictions
}
\description{
This function will apply the selected threshold criterias to MaxEnt model projection(s) of a 'mcmp' object
and save on the folder "3_out.MaxEnt/Mdls.[species name]/Mdls.thrshld". For each projection (species and climatic
scenario), two layers will be generated, one with suitability above the threshold value and another with presence/absence only.
}
\examples{
mods.thrshld <- thr(mcmp=mxnt.mdls.preds, thrshld.i = 4:6, pred.args, path.mdls)
plot(mods.thrshld[[1]][[2]]) # continuous
plot(mods.thrshld[[2]][[2]]) # binary
}
\seealso{
\code{\link{thrB}}
}
