% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{bffr.batch}
\alias{bffr.batch}
\title{Create buffer based on species polygons}
\usage{
bffr.batch(occ.polys, bffr.width = NULL, mult = 0.2, quadsegs = 100,
  crs.set = NULL, plot = T)
}
\arguments{
\item{occ.polys}{list of SpatialPolygons, usually obj returned from poly.c.batch()}

\item{bffr.width}{Buffer width. See 'width' of ?rgeos::gBuffer}

\item{mult}{How much expand bffr.width}

\item{quadsegs}{Number of line segments to use to approximate a quarter circle.}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}

\item{plot}{Boolean, to draw plots or not}
}
\value{
A named list of SpatialPolygons
}
\description{
This funcion will create a buffer using species polygons. Buffer width may be manually specified or
calculated based on the extent of the SpatialPolygons object (i.e. mean of latitudinal and longitudinal extent).
If width is calculated based on the extent of the SpatialPolygons object, it can be adjusted (enlarged or reduced)
using 'mult' argument.
}
\examples{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- poly.c.batch(spp.occ.list)
occ.b <- bffr.batch(occ.polys, bffr.width=1.5)
}
\seealso{
\code{\link[rgeos]{gBuffer}}, \code{\link{poly.c.batch}}
}
