% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{bind.shp}
\alias{bind.shp}
\title{Bind list of SpatialPolygons into a single SpatialPolygon object}
\usage{
bind.shp(occ.polys, sp.nm = "sp.nm", save = T,
  crs.set = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{occ.polys}{list of SpatialPolygons to bind}

\item{sp.nm}{Species name, used on saving shapefile}

\item{save}{Should save polygons on disk?}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\value{
shapefile with binded polygons
}
\description{
Bind list of SpatialPolygons into a single SpatialPolygon object
}
