% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{bind.shp}
\alias{bind.shp}
\title{Bind list of SpatialPolygons into a single SpatialPolygons object}
\usage{
bind.shp(occ.polys, sp.nm = "sp.nm", save = TRUE,
  crs.set = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{occ.polys}{list of SpatialPolygons to bind}

\item{sp.nm}{Species name, used on saving shapefile}

\item{save}{Should save polygons on disk?}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\value{
shapefile with binded polygons
}
\description{
This function will bind a list of splitted polygons (from polySplt()) into a single SpatialPolygons object.
}
\seealso{
\code{\link{polyCB}}, \code{\link{polyC}}, \code{\link{polySplt}}, \code{\link[NbClust]{NbClust}}
}
\keyword{internal}
