% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.f.consensus_projections.R
\name{consensus_scn}
\alias{consensus_scn}
\title{Create consensual projections of climatic scenarios}
\usage{
consensus_scn(mcm, groups, ref = NULL, sp.nm = "species", save = T)
}
\arguments{
\item{ref}{Object returned by \code{\link{proj_mdl_b}}, containing a list of calibrated models
and model projections for each species.}

\item{sp.nm}{Species name. Used to name the output folder}
}
\value{
A 'mcmp.l' object. An object returned from function \code{\link{proj_mdl_b}},
 containing the consensual projections for each element (species) of the list
}
\description{
This function creates consensual projections of climatic scenarios
(e.g. GCMs, RCPs, SSPs).
User need to supply a 'mcmp' object (returned by \code{\link{proj_mdl}}),
and a list containing names for grouping projections.
Projections will be grouped by matching names in the list of projections
}
\seealso{
\code{\link{consensus_scn_b}}
}
