% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.f.tuning.R
\name{ENMevaluate.batch}
\alias{ENMevaluate.batch}
\title{Tuning and evaluation of ENMs with Maxent for several species using ENMeval}
\usage{
ENMevaluate.batch(occ.l, a.calib.l, bg.coords.l = NULL, occ.grp = NULL,
  bg.grp = NULL, RMvalues = seq(0.5, 4, 0.5), fc = c("L", "LQ", "H",
  "LQH", "LQHP", "LQHPT"), categoricals = NULL, n.bg = 10000,
  method = NULL, overlap = FALSE, aggregation.factor = c(2, 2),
  kfolds = NA, bin.output = FALSE, clamp = TRUE, rasterPreds = TRUE,
  parallel = FALSE, numCores = NULL, progbar = TRUE,
  updateProgress = FALSE, ...)
}
\arguments{
\item{occ.l}{list of species occurrence data}

\item{a.calib.l}{list of predictors (cropped environmental variables) for model tuning. Used in model calibration. Argument 'x' of dismo::maxent. Raster* object or SpatialGridDataFrame, containing grids with
predictor variables. These will be used to extract values from for the point locations. Can
also be a data.frame, in which case each column should be a predictor variable and each row
a presence or background record.}

\item{bg.coords.l}{list of background localities. Two-column matrix or data.frame of longitude and latitude (in that order) of background localities (required for 'user' method).}

\item{occ.grp}{ Vector of bins of occurrence localities (required for '\code{user}' method). }

\item{bg.grp}{ Vector of bins of background localities (required for '\code{user}' method). }

\item{RMvalues}{ Vector of (non-negative) values to use for the regularization multiplier. }

\item{fc}{ Character vector of feature class combinations to be included in analysis. }

\item{categoricals}{ Vector indicating which (if any) of the input environmental layers are categorical. }

\item{n.bg}{ The number of random background localities to draw from the study extent. }

\item{method}{ Character string designating the method used for data partitioning.  Choices are: \code{"jackknife", "randomkfold", "user", "block", "checkerboard1", "checkerboard2"}.  See details and \code{\link{get.evaluation.bins}} for more information. }

\item{overlap}{ logical; If \code{TRUE}, provides pairwise metric of niche overlap (see details and \code{\link{calc.niche.overlap}}). }

\item{aggregation.factor}{ List giving the factor by which the original input grid should be aggregated for checkerboard partitioning methods (see details and \code{\link{get.evaluation.bins}}). }

\item{kfolds}{ Number of bins to use in the \emph{k}-fold random method of data partitioning. }

\item{bin.output}{ logical; If \code{TRUE}, appends evaluations metrics for each evaluation bin to results table (i.e., in addition to the average values across bins). }

\item{clamp}{ logical; If \code{TRUE}, 'clamping' is used (see Maxent documentation and tutorial for more details). }

\item{rasterPreds}{ logical; If \code{TRUE}, the \code{predict} function from \code{dismo} is used to predict each full model across the extent of the input environmental variables.  Note that AICc (and associated values) are \emph{NOT} calculated if \code{rasterPreds=FALSE} because these calculations require the predicted surfaces.  However, setting to \code{FALSE} can significantly reduce run time.}

\item{parallel}{ logical; If \code{TRUE}, parallel processing is used to execute tuning function. }

\item{numCores}{ numeric; indicates the number of cores to use if running in parallel. If \code{parallel=TRUE} and this is not specified, the total number of available cores are used.}

\item{progbar}{  logical; used internally. }

\item{updateProgress}{ logical; used internally. }

\item{...}{ character vector; use this to pass other arguments (e.g., prevalence) to the `maxent` call.  Note that not all options are functional or relevant.}
}
\description{
This function is a wrapper for ENMeval::ENMevaluate. See ?ENMeval::ENMevaluate for details
It works with a named list of species occurrence data (occ.l) and a list of
cropped environmental variables (a.calib.l) for model tuning.
}
\examples{
ENMeval.res.lst <- ENMevaluate.batch(occ.locs, occ.b.env, parallel = T , numCores = 7)
}
