% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{env.cut}
\alias{env.cut}
\title{Cut calibration area based on a list of SpatialPolygons
Use a list of SpatialPolygons to crop environmental variables for each species.}
\usage{
env.cut(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Cut calibration area based on a list of SpatialPolygons
Use a list of SpatialPolygons to crop environmental variables for each species.
}
