% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1b.f.occ.thin.R
\name{env_thin}
\alias{env_thin}
\title{Thinning of occurrence records using environmental information}
\usage{
env_thin(
  p,
  predictors,
  long.col = NULL,
  lat.col = NULL,
  bins = 20,
  file = NULL,
  plot = F,
  verbose = F
)
}
\arguments{
\item{p}{Two column matrix or data.frame with point coordinates
or \code{\link[sp]{SpatialPoints}} of occurrence records.}

\item{predictors}{\code{\link[raster]{Raster-class}} object of environmental
predictor variables.}

\item{lat.col, long.col}{Name of columns that contain coordinates (latitude and longitude)}

\item{bins}{Number of bins to divide each environmental variable.}

\item{file}{either a character string naming a file or a connection open for writing.}

\item{plot}{Logical. Should results be plotted?}

\item{verbose}{Logical. Should information be printed on screen?}
}
\value{
A data.frame, matrix, sp or sf object containing the selected records.
}
\description{
The function thins spatial points to minimize biases in the environmental space.
It divides de range of variables into a desired number of @param bins and selects points
closer to the bin center.
}
\examples{
\dontrun{
file <- paste(system.file(package="dismo"), "/ex/bradypus.csv", sep="")
bradypus <- read.table(file, header=TRUE, sep=',')
coord <- bradypus [,2:3]

# Download data for present
dir.create("rasters")
library(raster)
predictors <- getData('worldclim', var='bio', res=10, path="rasters")

#cor(data)
#vars <- ENMwizard::select_vars(cutoff = 0.9, corr.mat = cor(data), names.only = T)[[1]]
#length(vars)
env_thin(coord, predictors)
}

}
\seealso{
\code{\link{env_thin_b}}, \code{\link{load_env_thin_occ}}
}
