% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.f.tuning.final.R
\name{f.args}
\alias{f.args}
\title{Creating arguments for selected models}
\usage{
f.args(x, mSel = c("AvgAIC", "LowAIC", "OR", "AUC"), wAICsum = 0.99,
  save = "B", randomseed = FALSE, responsecurves = TRUE,
  arg1 = "noaddsamplestobackground", arg2 = "noautofeature")
}
\arguments{
\item{x}{Slot "results" of object of class ENMevaluation}

\item{mSel}{character vector. Which criteria to use when selecting model(s). Currently implemented:
"AvgAIC", "LowAIC", "OR", "AUC"}

\item{wAICsum}{cumulative sum of top ranked models for which arguments will be created}

\item{save}{should save args only ("A"), selected models only ("M") or both ("B")?}

\item{randomseed}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{responsecurves}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg1}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg2}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}
}
\value{
A vector of args (if save="A"), data.frame of selected models (if save="M") or
a list with both, args and selected models, (if save="B")
}
\description{
This function will read an object of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return the necessary arguments for final model calibration and predictions.
}
\seealso{
\code{\link{mxnt.c}}, \code{\link{mxnt.c.batch}}, \code{\link[dismo]{maxent}}, \code{\link[ENMeval]{ENMevaluate}},
\code{\link{mxnt.p}}, \code{\link{mxnt.p.batch.mscn}}
#' @examples
ENMeval.res.lst <- ENMevaluate.batch(occ.locs, occ.b.env, parallel = T , numCores = 7)
f.args(ENMeval.res.lst[[1]]@results)
}
\keyword{internal}
