% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{f.args}
\alias{f.args}
\title{Model selection and creation of MaxEnt arguments for selected models}
\usage{
f.args(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will read an object of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return the results table with models selected by the chosen criteria. It can also return the
necessary arguments for final model calibration and predictions.
}
