% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.f.metrics.R
\name{f.FPA}
\alias{f.FPA}
\title{Compute "Fractional predicted area" ('n of occupied pixels'/n) for multiple scenarios}
\usage{
f.FPA(mtp.l, digits = 3)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' FPAs computed for each climatic scenario, threshold and model criteria
}
\description{
Compute "Fractional predicted area" ('n of occupied pixels'/total n) or ('area of occupied pixels'/total area)
}
\examples{
f.FPA(mtp.l=mods.thrshld.lst)
}
\seealso{
\code{\link{f.area.occ.mscn}}, \code{\link{f.var.ci}}, \code{\link{f.OR}}, \code{\link{f.raster.overlap.mscn}}
}
