% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.f.metrics.R
\name{f.FPA}
\alias{f.FPA}
\title{Compute "Fractional predicted area" ('n of occupied pixels'/n) for multiple scenarios}
\usage{
f.FPA(mtp.l, digits = 3)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' FPAs computed for each climatic scenario, threshold and model criteria
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
f.FPA.mscn(mtp.l)
}
