% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.f.metrics.R
\name{f.OR}
\alias{f.OR}
\title{Compute "Fractional predicted area" ('n of occupied pixels'/n)}
\usage{
f.OR(mtp.l, occ.l, current.pred.nm = "current", digits = 3)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{occ.l}{list of species occurrence data.}

\item{current.pred.nm}{name to locate climatic scenario (usually "current") used to calibrate maxent models}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' ORs computed for the selected (current) climatic scenario and
each threshold and model criteria
}
\description{
Compute "Fractional predicted area" ('n of occupied pixels'/total n) or  ('area of occupied pixels'/total area)
}
\examples{
f.OR(mods.thrshld.lst, occ.locs, "current")
}
