% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.f.metrics.R
\name{f.OR}
\alias{f.OR}
\title{Compute "Omission Rate"}
\usage{
f.OR(mtp.l, occ.l, current.pred.nm = "current", digits = 3)
}
\arguments{
\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{occ.l}{list of species occurrence data.}

\item{current.pred.nm}{name to locate climatic scenario (usually "current") used to calibrate maxent models}

\item{digits}{integer indicating the number of decimal places. see ?round for details.}
}
\value{
A list of species' ORs computed for the selected (current) climatic scenario and
each threshold and model criteria
}
\description{
Compute "Omission Rate" of species occurence points for a climatic scenario (usually "current")
}
\examples{
f.OR(mods.thrshld.lst, occ.locs, "current")
}
\seealso{
\code{\link{f.area.occ.mscn}}, \code{\link{f.var.ci}}, \code{\link{f.FPA}}, \code{\link{f.raster.overlap.mscn}}
}
