% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{f.plot.mxnt.preds}
\alias{f.plot.mxnt.preds}
\title{Plot differences among model predictions selected from several criteria}
\usage{
f.plot.mxnt.preds(mmp.spl, mtp.spl, basemap, pred.nm = "")
}
\arguments{
\item{mmp.spl}{List of stack or brick of predictions to apply the threshold}

\item{mtp.spl}{List of stack or brick of thresholded predictions}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}

\item{pred.nm}{name of prediction to be appended to the final name. Usually "pres", "past" or "fut".}
}
\value{
won't return any object. Will save pdf's with differences among model predictions
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
f.plot.mxnt.preds(mxnt.mdls.preds.lst, mods.thrshld.lst, basemap=NewWorld)
}
