% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.f.threshold.R
\name{f.thr.batch}
\alias{f.thr.batch}
\title{Function Title (short description)}
\usage{
f.thr.batch(mxnt.mdls.preds.splst, pred.nm = "", thrshld.i = 4:6)
}
\arguments{
\item{mxnt.mdls.preds.splst}{List of stack/brick of predictions to apply the threshold}

\item{pred.nm}{name of prediction to... TODO}

\item{thrshld.i}{List of threshold criteria to be applied}
}
\value{
objects returned from function
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
plot(mxnt.mdls.preds.lst[[1]][[4]]) # MaxEnt predictions, based on the model selection criteria
}
