% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{f.thr.batch}
\alias{f.thr.batch}
\title{Apply threshold for MaxEnt projections for multiple species}
\usage{
f.thr.batch(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will apply the selected threshold criterias to MaxEnt model projection(s) of a 'mcmp.l' object
and save on the folder "3_out.MaxEnt/Mdls.[species name]/Mdls.thrshld". For each projection (species and climatic
scenario), two layers will be generated, one with suitability above the threshold value and another with presence/absence only.
}
