% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.f.threshold.R
\name{f.thr.batch}
\alias{f.thr.batch}
\title{Apply threshold for all predictions}
\usage{
f.thr.batch(mmp.spl, pred.nm = "", thrshld.i = 4:6)
}
\arguments{
\item{mmp.spl}{List of stack or brick of predictions to apply the threshold}

\item{pred.nm}{name of prediction to be appended to the final name. Usually "pres", "past" or "fut".}

\item{thrshld.i}{List of threshold criteria to be applied}
}
\value{
List of stack or brick of thresholded predictions
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
mods.thrshld.lst <- f.thr.batch(mxnt.mdls.preds.pf)
}
