% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.f.threshold.R
\name{f.thr}
\alias{f.thr}
\title{Apply threshold for a prediction}
\usage{
f.thr(mmp.spi, pred.nm = "", thrshld.i = 4:6)
}
\arguments{
\item{mmp.spi}{Stack or brick of predictions to apply the threshold}

\item{pred.nm}{name of prediction to be appended to the final name. Usually "pres", "past" or "fut".}

\item{thrshld.i}{List of threshold criteria to be applied}
}
\value{
Stack or brick of thresholded predictions
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
mods.thrshld <- f.thr(mxnt.mdls.preds, thrshld.i = 4:6, pred.args, path.mdls)
plot(mods.thrshld[[1]][[2]]) # continuous
plot(mods.thrshld[[2]][[2]]) # binary
}
