% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.f.threshold.R
\name{f.thr}
\alias{f.thr}
\title{Apply threshold for a prediction}
\usage{
f.thr(mcmp.spi, scn.nm = "", thrshld.i = 4:6, path.mdls = NULL)
}
\arguments{
\item{mcmp.spi}{Species "i" of a object returned by "mxnt.p.batch.mscn", containing a list of
calibrated models and model projections for each species}

\item{scn.nm}{Name of climatic scenario to be looked for}

\item{thrshld.i}{List of threshold criteria to be applied. Use numbers to choose the desired one(s). Current options:
1 - Fixed.cumulative.value.1 (fcv1),
2 - Fixed.cumulative.value.5 (fcv5),
3 - Fixed.cumulative.value.10 (fcv10),
4 - Minimum.training.presence (mtp),
5 - 10.percentile.training.presence (x10ptp),
6 - Equal.training.sensitivity.and.specificity (etss),
7 - Maximum.training.sensitivity.plus.specificity (mtss),
8 - Balance.training.omission.predicted.area.and.threshold.value (bto),
9 - Equate.entropy.of.thresholded.and.original.distributions (eetd)}

\item{path.mdls}{Path where thresholded rasters will be saved}
}
\value{
Stack or brick of thresholded predictions
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
mods.thrshld <- f.thr(mxnt.mdls.preds, thrshld.i = 4:6, pred.args, path.mdls)
plot(mods.thrshld[[1]][[2]]) # continuous
plot(mods.thrshld[[2]][[2]]) # binary
}
