% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.f.metrics.R
\name{f.var.ci}
\alias{f.var.ci}
\title{Compute variable contribution and importance}
\usage{
f.var.ci(mcmp.l)
}
\arguments{
\item{mcmp.l}{Object returned by "mxnt.p.batch.mscn", containing a list of calibrated models
and model projections for each species.}
}
\value{
List of arrays containing variable contribution and importance for each species
}
\description{
Compute variable contribution and importance for each model
}
\examples{
f.var.ci(mcmp.l = mxnt.mdls.preds.lst)
}
\seealso{
\code{\link[dismo]{maxent}}, \code{\link{f.area.occ.mscn}}, \code{\link{f.OR}}, \code{\link{f.FPA}}, \code{\link{f.raster.overlap.mscn}}
}
