% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1.f.data.prep.R
\name{loadTocc}
\alias{loadTocc}
\title{Load "thin.batch" filtered occurrence data}
\usage{
loadTocc(occ.list.thin, from.disk = FALSE)
}
\arguments{
\item{occ.list.thin}{named list returned from "thin.batch"}

\item{from.disk}{boolean. Read from disk or from one of thinned datasets stored in 'occ.list.thin' obj}
}
\value{
named list of thinned occurence data for each species in the list
}
\description{
Load "thin.batch" filtered occurrence data
}
\examples{
occ.locs <- loadTocc(thinned.dataset.batch)
}
\seealso{
\code{\link[spThin]{thin}}, \code{\link{thin.batch}}
}
