% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{loadTocc}
\alias{loadTocc}
\title{Load filtered occurrence data}
\usage{
loadTocc(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Load filtered occurrence data from object returned by "thin_b" function
}
