% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{make.underscript}
\alias{make.underscript}
\title{Format (underscript) selected texts for plotting}
\usage{
make.underscript(x)
}
\arguments{
\item{x}{list of text to be formatted}
}
\value{
list of formatted text
}
\description{
Format (underscript) selected texts (criteria used to select models) to be used on plotting.
}
\examples{
make.underscript(c("AUC (OR10p)", "AUC (ORlpt)", "OR10p (AUC)", "ORlpt (AUC)"))
}
\seealso{
\code{\link{f.plot.mxnt.preds}}, \code{\link{f.plot.mxnt.preds.mscn}}, \code{\link{f.plot.scn.diff}}
}
\keyword{internal}

