% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{make.underscript}
\alias{make.underscript}
\title{Format (underscript) selected texts for plotting}
\usage{
make.underscript(x)
}
\arguments{
\item{x}{list of text to be formatted}
}
\value{
list of formatted text
}
\description{
General function description. A short paragraph (or more) describing what the function does.
}
\examples{
make.underscript(c("AUC (OR10p)", "AUC (ORlpt)", "OR10p (AUC)", "ORlpt (AUC)"))
}
