% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.f.calibration.R
\name{mod_sel}
\alias{mod_sel}
\title{Model selection and creation of MaxEnt arguments for selected models}
\usage{
mod_sel(
  x,
  mSel = c("AvgAIC", "EBPM", "WAAUC", "ESOR", "LowAIC", "OR", "AUC"),
  wAICsum = 0.99,
  dAICc = 2,
  randomseed = FALSE,
  responsecurves = TRUE,
  arg1 = "noaddsamplestobackground",
  arg2 = "noautofeature",
  save = "M"
)
}
\arguments{
\item{x}{Object of class ENMevaluation}

\item{mSel}{Character vector. Which criteria to use when selecting model(s). Currently implemented:
"AvgAIC", "LowAIC", "OR", "AUC"}

\item{wAICsum}{Cumulative sum of top ranked models for which arguments will be created}

\item{dAICc}{Maximum delta AICc of models to be selected.}

\item{randomseed}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{responsecurves}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg1}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg2}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{save}{Should save args only ("A"), selected models only ("M") or both ("B")?}
}
\value{
A vector of args (if save="A"), data.frame of selected models (if save="M") or
a list with both, args and selected models, (if save="B")
}
\description{
This function will read an object of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return the results table with models selected by the chosen criteria. It can also return the
necessary arguments for final model calibration and predictions.
}
\examples{
\dontrun{
ENMeval.res.lst <- ENMevaluate(occ.locs, occ.b.env, parallel = F , numCores = 1)
mod_sel(ENMeval.res.lst[[1]])
}
}
\seealso{
\code{\link{calib_mdl}}, \code{\link{calib_mdl_b}}, \code{\link[dismo]{maxent}}, \code{\link[ENMeval]{ENMevaluate}},
\code{\link{proj_mdl}}, \code{\link{proj_mdl_b}}
}
