% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.3.f.metrics.MOP.R
\name{mop}
\alias{mop}
\title{Extrapolation risk analysis}
\usage{
mop(M, G, p = 0.1, q = 0.1, min.M.sz = 100, filename = NULL,
  scn.nm = "", numCores = 1, ...)
}
\arguments{
\item{M}{RasterStack of environmental variables from calibration area}

\item{G}{RasterStack of environmental variables from projection area}

\item{p}{Percent of values, sampled from calibration area, used as
reference to calculate the MOP. Must be >0 and <=1.}

\item{q}{Quantile. Proportion of closest points in M is to be compared
with G to calculate the MOP. Must be >0 and <=1.}

\item{min.M.sz}{Threshold value to be used to compute OR}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
Data frame with average and variance of OR values across partition groups of data
}
\description{
This function will compute the omission rate (OR) for a species' AICc Averaged Model
from a 'mcmp' object, based on the selected threshold value.
}
\seealso{
\code{\link{mop_b}}
}
