% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.f.tuning.final.R
\name{mxnt.c}
\alias{mxnt.c}
\title{Calibrating selected models}
\usage{
mxnt.c(ENMeval.o, sp.nm, a.calib, occ = NULL, use.ENMeval.bgpts = TRUE,
  nbg = 10000, formt = "raster", pred.args = c("outputformat=cloglog",
  "doclamp=true", "pictures=true"), wAICsum = 0.99, randomseed = FALSE,
  responsecurves = TRUE, arg1 = "noaddsamplestobackground",
  arg2 = "noautofeature", numCores = 1, parallelTunning = TRUE)
}
\arguments{
\item{ENMeval.o}{Object of class ENMevaluation}

\item{sp.nm}{Species name. Used to name the output folder}

\item{a.calib}{Predictors (cropped environmental variables) for model tuning. Used in model calibration. Argument 'x' of dismo::maxent. Raster* object or SpatialGridDataFrame, containing grids with
predictor variables. These will be used to extract values from for the point locations. Can
also be a data.frame, in which case each column should be a predictor variable and each row
a presence or background record.}

\item{occ}{Occurrence data. Argument 'p' of dismo::maxent. This can be a data.frame, matrix,
SpatialPoints* object, or a vector. If p is a data.frame or matrix it represents a set of point
locations; and it must have two columns with the first being the x-coordinate (longitude) and
the second the y-coordinate (latitude). Coordinates can also be specified with a SpatialPoints*
object
If x is a data.frame, p should be a vector with a length equal to nrow(x) and contain 0
(background) and 1 (presence) values, to indicate which records (rows) in data.frame x are
presence records, and which are background records}

\item{use.ENMeval.bgpts}{Logical. Use background points from ENMeval or sample new ones?}

\item{nbg}{Number of background points to use. These are sampled randomly from the cells that are not \code{NA} in the first predictor variable. Ignored if background points are specified with argument \code{a}}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{pred.args}{Charater. Argument 'args' of dismo::maxent. Additional argument that can be
passed to MaxEnt. See the MaxEnt help for more information. The R maxent function only uses the
arguments relevant to model fitting. There is no point in using args='outputformat=raw' when
*fitting* the model; but you can use arguments relevant for *prediction* when using the predict
function. Some other arguments do not apply at all to the R implementation. An example is
'outputfiletype', because the 'predict' function has its own 'filename' argument for that.}

\item{wAICsum}{cumulative sum of top ranked models for which arguments will be created}

\item{randomseed}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{responsecurves}{logical. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg1}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{arg2}{charater. Args to be passed to dismo::maxent. See ?dismo::maxent and the MaxEnt help for more information.}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{parallelTunning}{Should parallelize within species (parallelTunning=TRUE) or between species (parallelTunning=FALSE)}
}
\value{
A 'mcm' (mxnt.c.mdls, Maxent Calibrated Models). A list containing the models ('selected.mdls') used for model calibration,
calibrated maxent models ('mxnt.mdls'), and arguments used for calibration ('pred.args').
}
\description{
This function will read an object of class ENMevaluation (See ?ENMeval::ENMevaluate for details) and
return selected maxent models calibrated.
}
\seealso{
\code{\link{f.args}}, \code{\link{mxnt.c.batch}}, \code{\link[dismo]{maxent}}, \code{\link[ENMeval]{ENMevaluate}},
\code{\link{mxnt.p}}, \code{\link{mxnt.p.batch.mscn}}
}
