% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.f.mscn.projection.R
\name{mxnt.p.batch.mscn}
\alias{mxnt.p.batch.mscn}
\title{Projecting calibrated MaxEnt models for several species onto multiple environmental scenarios}
\usage{
mxnt.p.batch.mscn(mxnt.c.mdls.lst, a.proj.l, formt = "raster", numCores = 1)
}
\arguments{
\item{mxnt.c.mdls.lst}{A list of objects returned by "mxnt.cp", containing calibrated models.}

\item{a.proj.l}{A list of Raster* objects or data.frames where models will be projected. Argument 'x' of dismo::predict}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
A list of objects returned from function "mxnt.p", containing the new (multiple) projections for each element (species) of the list
}
\description{
This function will read an object returned by "mxnt.cp.batch", read the calibrated models and project into
several environmental (areas/climatic) scenarios (specified in a.proj.l). These new projections will be returned together with (appended to)
each element (species) the original object.
}
\examples{
mxnt.mdls.preds.pf <- mxnt.p.batch.Mscn(mxnt.mdls.preds.lst, a.proj.l = area.projection.pf)
}
