% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{mxnt.p.batch.mscn}
\alias{mxnt.p.batch.mscn}
\title{Project calibrated MaxEnt models for several species onto multiple environmental scenarios}
\usage{
mxnt.p.batch.mscn(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will read an object returned by "calib_mdl_b", read the calibrated models and project into
several environmental (areas/climatic) scenarios (specified in a.proj.l). These new projections will be returned together with (appended to)
each element (species) the original object.
}
