% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.f.mscn.projection.BKP.R
\name{mxnt.p.batch}
\alias{mxnt.p.batch}
\title{Projecting calibrated MaxEnt models for several species}
\usage{
mxnt.p.batch(mxnt.c.mdls.lst, pred.nm = "fut", a.proj.l, formt = "raster",
  numCores = 1)
}
\arguments{
\item{mxnt.c.mdls.lst}{A list of objects returned by "mxnt.cp", containing calibrated models.}

\item{pred.nm}{Character. Prefix to add to projection name (e.g. "fut" or "past")}

\item{a.proj.l}{A list of Raster* objects or data.frames where models will be projected. Argument 'x' of dismo::predict}

\item{formt}{Character. Output file type. Argument 'format' of raster::writeRaster}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
A list of objects returned from function "mxnt.p"
}
\description{
This function will read an object returned by "mxnt.cp.batch", read the calibrated models and project into
new areas/climatic scenarios. These new projections will be returned together with (appended to)
each element (species) the original object
}
\examples{
mxnt.mdls.preds <- mxnt.p.batch(mxnt.c.mdls.lst = mxnt.mdls.preds.lst[1],
}
