% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{plotMdlDiff}
\alias{plotMdlDiff}
\title{Plot differences between predictions of models selected from several criteria}
\usage{
plotMdlDiff(mcmp.l, mtp.l, basemap = NULL, save = FALSE, numCores = 1)
}
\arguments{
\item{mcmp.l}{Object returned by "mxnt.p.batch.mscn", containing a list of calibrated models
and model projections for each species.}

\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}

\item{save}{Logical. If TRUE will save plots in pdf.}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}

\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}
}
\value{
won't return any object. Will save pdf's with differences among model predictions

won't return any object. Will save pdf's with differences among model predictions (for multiple climatic scenarios)
}
\description{
Plot differences among predictions of species model projetions selected using distinct criteria.
This function will save the figures on pdf files in the folder "Mdls.thrshld/figs".
}
\examples{
f.plot.mxnt.preds(mxnt.mdls.preds.lst, mods.thrshld.lst, basemap=NewWorld)
plotMdlDiff(mxnt.mdls.preds.lst, mods.thrshld.lst, basemap=NewWorld)
plotMdlDiff(mxnt.mdls.preds.pf[1], mods.thrshld.lst[1], basemap=NewWorld)
}
\seealso{
\code{\link{plotMdlDiff}}, \code{\link{plotScnDiff}}

\code{\link{plotScnDiff}}
}
\keyword{"AUC")}
\keyword{"AvgAIC",}
\keyword{"LowAIC",}
\keyword{"Mdls.thrshld/figs".}
\keyword{"OR",}
\keyword{(e.g.}
\keyword{(for}
\keyword{Plot}
\keyword{This}
\keyword{among}
\keyword{climatic}
\keyword{criteria}
\keyword{criteria.}
\keyword{differences}
\keyword{distinct}
\keyword{figures}
\keyword{files}
\keyword{folder}
\keyword{from}
\keyword{function}
\keyword{in}
\keyword{internal}
\keyword{model}
\keyword{multiple}
\keyword{on}
\keyword{pdf}
\keyword{predictions}
\keyword{save}
\keyword{scenarios)}
\keyword{selected}
\keyword{several}
\keyword{the}
\keyword{using}
\keyword{will}
