% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.f.plot.mdl.sel.crit.diff.R
\name{plotScnDiff}
\alias{plotScnDiff}
\title{Plot differences between climatic scenarios}
\usage{
plotScnDiff(mcmp.l, mtp.l, mSel = "LowAIC", sel.clim.scn = "current",
  basemap = NULL, save = FALSE, numCores = 1)
}
\arguments{
\item{mcmp.l}{Object returned by "mxnt.p.batch.mscn", containing a list of calibrated models
and model projections for each species.}

\item{mtp.l}{List of stack or brick of thresholded predictions}

\item{mSel}{Name of selection criteria to be compared: AvgAICc, LowAICc, avg.test.AUC10pct, avg.test.AUC.MTP,
avg.test.or10pct, avg.test.orMTP}

\item{sel.clim.scn}{Selected climatic scenario to compare with all others. Usually "current" one.}

\item{basemap}{Shapefile to be plotted with. Usually a continent or country shapefile}

\item{save}{Export to pdf or not?}

\item{numCores}{Number of cores to use for parallelization. If set to 1, no paralellization is performed}
}
\value{
won't return any object. Will save pdf's with differences among model predictions (for multiple climatic scenarios)
}
\description{
Plot differences between a selected climatic scenario and all other climatic scenarios for each species.
This function will plota and (optionally) save the figures on pdf files in the folder "Mdls.thrshld/figs".
}
\examples{
plotScnDiff(mcmp.l=mxnt.mdls.preds.cf, mtp.l=mods.thrshld.lst)
}
\seealso{
\code{\link{plotMdlDiff}}
}
