% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{poly.c.batch}
\alias{poly.c.batch}
\title{Create minimum convex polygon based on coordinates of species occurence data for several species}
\usage{
poly.c.batch(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will use a list of coordinates of species occurence data and create minimum convex polygons
for each element in the list.
It is possible to create concave or convex polygons, create several small polygons based on clusters
of points.
}
