% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{poly.c.batch}
\alias{poly.c.batch}
\title{Create occ polygon for several species}
\usage{
poly.c.batch(spp.occ.list, convex = T, alpha = 10, plot = T,
  crs.set = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{spp.occ.list}{A named list of spatial data.frame of species occurence points}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{Value of \eqn{\alpha}.}

\item{plot}{logical. Plot results or not?}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\value{
A named list of spatial polygons built using coordinates
}
\description{
Create occ polygon for several species
}
\examples{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- poly.c.batch(spp.occ.list)
occ.polys <- poly.c.batch(spp.occ.list, convex=T, alpha=10)
}
