% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{poly.c.batch}
\alias{poly.c.batch}
\title{Create occ polygon for several species}
\usage{
poly.c.batch(spp.occ.list, crs.set = NULL, convex = T, alpha = 10,
  plot = T)
}
\arguments{
\item{spp.occ.list}{A named list of spatial data.frame of species occurence points}

\item{crs.set}{crs value used ???}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{see ?alphahull::ashape}

\item{plot}{logical. Plot results or not?}
}
\value{
A named list of spatial polygons built using coordinates
}
\description{
Create occ polygon for several species
}
\examples{
occ_polys <- poly.c.batch(spp.occ.list, convex=T, alpha=10)
}
