% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{poly.c}
\alias{poly.c}
\title{Create minimum convex polygon based on species occurence data}
\usage{
poly.c(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function will create minimum convex polygon based on coordinates of species occurence data.
}
