% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{poly.c}
\alias{poly.c}
\title{Create polygon based on species occurence coordinates}
\usage{
poly.c(occ.spdf, lr.nm = "sp_nm", convex = T, alpha = 10,
  crs.set = NULL)
}
\arguments{
\item{occ.spdf}{A spatial data.frame of coordinates, usually species occurence coordinates}

\item{lr.nm}{Polygon output name}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{see ?alphahull::ashape}

\item{crs.set}{crs value used ???}
}
\value{
spatial polygon of occurencies built using coordinates
}
\description{
Create polygon based on species occurence coordinates
}
\examples{
occ_poly <- poly.c(occ.spdf, lr.nm="occ_poly")
plot(occ_poly)
}
