% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{poly.splt}
\alias{poly.splt}
\title{split a species occ polygon (whenever distribution seems disjoint) into K polygons and save in a single .shp}
\usage{
poly.splt(spp.occ, k = 2, convex = T, alpha = 10, sp.nm = "sp1",
  crs.set = NULL)
}
\arguments{
\item{spp.occ}{species occurence coordinates}

\item{k}{number of polygons to create based on coordinates}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{see ?alphahull::ashape}

\item{sp.nm}{name used to save file}

\item{crs.set}{crs value used ???}
}
\value{
spatial polygons built using coordinates
}
\description{
split a species occ polygon (whenever distribution seems disjoint) into K polygons and save in a single .shp
}
\examples{
occ_polys$Bvarieg <- poly.splt(spp.occ = Bvarieg.occ, k=5)
}
