% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.f.data.prep.R
\name{poly.splt}
\alias{poly.splt}
\title{split a species occ polygon (whenever distribution seems disjoint) into K polygons and save in a single .shp}
\usage{
poly.splt(occ.spdf, k = NULL, c.m = "AP", convex = T, alpha = 10,
  sp.nm = "sp.nm", save = T, crs.set = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{occ.spdf}{species occurence coordinates}

\item{k}{number of polygons to create based on coordinates}

\item{c.m}{clustering method to find the best number of clusters (k). Currently E (Elbow) or (Affinity Propagation).}

\item{convex}{Concave or convex polygon (T or F)}

\item{alpha}{Value of \eqn{\alpha}.}

\item{sp.nm}{Species name, used on saving shapefile}

\item{save}{Should save polygons on disk?}

\item{crs.set}{set the coordinate reference system (CRS) of the polygons}
}
\value{
spatial polygons built using coordinates
}
\description{
split a species occ polygon (whenever distribution seems disjoint) into K polygons and save in a single .shp
}
\examples{
Bvarieg.occ <- read.table(paste(system.file(package="dismo"),
 "/ex/bradypus.csv", sep=""), header=TRUE, sep=",")
colnames(Bvarieg.occ) <- c("SPEC", "LONG", "LAT")
spp.occ.list <- list(Bvarieg = Bvarieg.occ)
occ.polys <- poly.c.batch(spp.occ.list)
occ.polys$Bvarieg <- poly.splt(occ.spdf = spp.occ.list$Bvarieg, k=5)
}
