% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{pred.a.batch.mscn}
\alias{pred.a.batch.mscn}
\title{Cut multiple projection areas (climatic scenarios) for multiple species (list of SpatialPolygons)}
\usage{
pred.a.batch.mscn(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function is a wrapper for "cut_projarea". See ?cut_projarea. This function delimits the projection area for
each of the species contained in the pred.polys named list and crops
multiple rasters/bricks (i.e. representing distinct climatic scenaries) based on the same criteria for each species.
}
