% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.a.batch.mscn}
\alias{pred.a.batch.mscn}
\title{Cut multiple projection areas (climatic scenarios) for multiple species (list of SpatialPolygons)}
\usage{
pred.a.batch.mscn(pred.polys, env.uncut.l, prj.nm = "", numCores = 1)
}
\arguments{
\item{pred.polys}{list of SpatialPolygons (usually of based on species occ points)
#' @seealso \code{\link{pred.a.poly}}, \code{\link{pred.a.poly.batch}},
\code{\link{pred.a}}, \code{\link{pred.a.batch.mscn}},
\code{\link{pred.a.rst}}, \code{\link{pred.a.batch.rst}}, \code{\link{pred.a.batch.rst.mscn}}}

\item{env.uncut.l}{list of raster/brick of environmental variables to be cut}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{numCores}{specify number of cores if aim run in parallel}
}
\value{
named list of cropped list of raster/brick of environmental variables
}
\description{
This function is a wrapper for "pred.a". See ?pred.a. This function delimitates the projection area for
each of the species contained in the pred.polys named list and crops
multiple rasters/bricks (i.e. representing distinct climatic scenaries) based on the same criteria for each species.
}
\seealso{
\code{\link{pred.a.poly}}, \code{\link{pred.a.poly.batch}},
\code{\link{pred.a}}, \code{\link{pred.a.batch}},
\code{\link{pred.a.rst}}, \code{\link{pred.a.batch.rst}}, \code{\link{pred.a.batch.rst.mscn}}
}
