% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.a.batch.rst.mscn}
\alias{pred.a.batch.rst.mscn}
\title{Cut multiple projection areas of multiple species based on a single SpatialPolygon object (e.g. Ecoregion)}
\usage{
pred.a.batch.rst.mscn(area.p, env.uncut.l, occ.polys, mask = FALSE,
  prj.nm = "", sp.nm = "mult.spp", numCores = 1)
}
\arguments{
\item{area.p}{SpatialPolygon to be used as reference to crop/mask environmental variables of all species}

\item{env.uncut.l}{list of raster/brick of environmental variables to be cut}

\item{occ.polys}{list of occ.poly SpatialPolygons.  see ?pred.a.poly for details.}

\item{mask}{Should mask raster? (i.e. only use area inside polygon. See ?raster::mask for details) or
use all spatial extent of area.p

#' @seealso \code{\link{pred.a.poly}}, \code{\link{pred.a.poly.batch}},
\code{\link{pred.a}}, \code{\link{pred.a.batch}}, \code{\link{pred.a.batch.mscn}},
\code{\link{pred.a.rst}}, \code{\link{pred.a.batch.rst.mscn}}}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{sp.nm}{name (of species) to give to saved object}

\item{numCores}{specify number of cores if aim run in parallel}
}
\value{
list of list with multiple environmental layers (raster/brick) cutted
}
\description{
This function will use a single SpatialPolygon to crop/mask multiple raster/brick objects
to be used on model projections.
}
\seealso{
\code{\link{pred.a.poly}}, \code{\link{pred.a.poly.batch}},
\code{\link{pred.a}}, \code{\link{pred.a.batch}}, \code{\link{pred.a.batch.mscn}},
\code{\link{pred.a.rst}}, \code{\link{pred.a.batch.rst}}
}
