% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{pred.a}
\alias{pred.a}
\title{Cut area for projection based on a list of SpatialPolygons}
\usage{
pred.a(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
This function is a wrapper for "cut_projarea". See ?cut_projarea
It works with a named list of pred.poly to delimit the projection area for each of the species.
}
