% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.a.rst}
\alias{pred.a.rst}
\title{Cut a projection area based on a SpatialPolygon (e.g. Ecoregion)}
\usage{
pred.a.rst(area_p, env_uncut, mask = F, prj.nm = "", sp.nm = "sp")
}
\arguments{
\item{area_p}{SpatialPolygon to be used as reference to crop/mask environmental variables of all species}

\item{env_uncut}{raster/brick of environmental variables to be cut}

\item{mask}{should use area_p to "mask" or "crop" env_uncut? See ?raster::mask and ?raster::crop for details}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}

\item{sp.nm}{name (of species) to give to saved object}
}
\value{
environmental layers (raster/brick) cutted
}
\description{
This function will use a single SpatialPolygon to crop/mask raster/brick objects to be used on model projections.
}
\examples{

}
