% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.f.proj.prep.R
\name{pred.ab}
\alias{pred.ab}
\title{Cut area for projection based on a list of SpatialPolygons}
\usage{
pred.ab(pred_polys, env_uncut, prj.nm = "")
}
\arguments{
\item{pred_polys}{list of SpatialPolygons (usually of based on species occ points)}

\item{env_uncut}{raster/brick of environmental variables to be cut}

\item{prj.nm}{climatic scenario name, usually "fut" or "past" is used to indicate a general timing of scenario}
}
\value{
named list of cropped raster or brick
}
\description{
This function is a wrapper for "pred.a". See ?pred.a
It works with a named list of pred_poly to delimitate the projection area for each of the species
}
\examples{
area_projection <- pred.ab(pred_polys, env_uncut, mult=.55, buffer=F)
plot(area_projection[[1]][[1]])
}
