% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx.Deprecated.functions.R
\name{thin.batch}
\alias{thin.batch}
\title{Spatially thin a list of species occurrence data}
\usage{
thin.batch(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Will use spThin optimisation algorithm to subset the dataset such that
all occurrence locations are a minimum distance apart. This process helps
reduce the effect of biases in observation records on the predictive
performance of ecological niche models.
}
\details{
Make sure coordinates are in decimal degrees. This function will use
great.circle.distance to thin the datasets
}
